fixes for cpp and c
