+++ examples/python/src/library.nim
@@ -1,0 +1,3 @@
+import nimview
+when isMainModule and system.appType != "lib" and not defined noMain:
+  start()
