+++ examples/python/src/nim.cfg
@@ -1,0 +1,1 @@
+--path:"../../../src/"
