+++ examples/svelte_todo/package.json
@@ -8,7 +8,7 @@
     "url": "git+https://github.com/marcomq/nimview.git"
   },
   "scripts": {
-    "build": "rollup -c && inliner -m --videos --preserve-comments dist/index.html > dist/nimview.html",
+    "build": "rollup -c && inliner -m --videos --preserve-comments dist/index.html > dist/inlined.html",
     "devbuild": "rollup -c",
     "autobuild": "rollup -c -w",
     "dev": "run-p start:dev autobuild",
