+++ src/js/nimview.cjs.js
@@ -1,0 +1,2 @@
+"use strict";function t(t,e,r,n,o,i,a){try{var u=t[i](a),s=u.value}catch(t){return void r(t)}u.done?e(s):Promise.resolve(s).then(n,o)}function e(e){return function(){var r=this,n=arguments;return new Promise((function(o,i){var a=e.apply(r,n);function u(e){t(a,o,i,u,s,"next",e)}function s(e){t(a,o,i,u,s,"throw",e)}u(void 0)}))}}function r(t){return(r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}var n={exports:{}},o=n.exports="undefined"!=typeof window&&window.Math==Math?window:"undefined"!=typeof self&&self.Math==Math?self:Function("return this")();"number"==typeof __g&&(__g=o);var i={}.hasOwnProperty,a=function(t,e){return i.call(t,e)},u=function(t){try{return!!t()}catch(t){return!0}},s=!u((function(){return 7!=Object.defineProperty({},"a",{get:function(){return 7}}).a})),c={exports:{}},f=c.exports={version:"2.6.12"};"number"==typeof __e&&(__e=f);var l={},h=function(t){return"object"==typeof t?null!==t:"function"==typeof t},p=h,y=function(t){if(!p(t))throw TypeError(t+" is not an object!");return t},d=h,v=n.exports.document,g=d(v)&&d(v.createElement),b=function(t){return g?v.createElement(t):{}},m=!s&&!u((function(){return 7!=Object.defineProperty(b("div"),"a",{get:function(){return 7}}).a})),w=h,x=function(t,e){if(!w(t))return t;var r,n;if(e&&"function"==typeof(r=t.toString)&&!w(n=r.call(t)))return n;if("function"==typeof(r=t.valueOf)&&!w(n=r.call(t)))return n;if(!e&&"function"==typeof(r=t.toString)&&!w(n=r.call(t)))return n;throw TypeError("Can't convert object to primitive value")},_=y,S=m,E=x,A=Object.defineProperty;l.f=s?Object.defineProperty:function(t,e,r){if(_(t),e=E(e,!0),_(r),S)try{return A(t,e,r)}catch(t){}if("get"in r||"set"in r)throw TypeError("Accessors not supported!");return"value"in r&&(t[e]=r.value),t};var O=function(t,e){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:e}},T=l,P=O,j=s?function(t,e,r){return T.f(t,e,P(1,r))}:function(t,e,r){return t[e]=r,t},I={exports:{}},L=0,F=Math.random(),R=function(t){return"Symbol(".concat(void 0===t?"":t,")_",(++L+F).toString(36))},N={exports:{}},k=c.exports,B=n.exports,M=B["__core-js_shared__"]||(B["__core-js_shared__"]={});(N.exports=function(t,e){return M[t]||(M[t]=void 0!==e?e:{})})("versions",[]).push({version:k.version,mode:"global",copyright:"© 2020 Denis Pushkarev (zloirock.ru)"});var U=N.exports("native-function-to-string",Function.toString),C=n.exports,D=j,G=a,q=R("src"),V=U,W=(""+V).split("toString");c.exports.inspectSource=function(t){return V.call(t)},(I.exports=function(t,e,r,n){var o="function"==typeof r;o&&(G(r,"name")||D(r,"name",e)),t[e]!==r&&(o&&(G(r,q)||D(r,q,t[e]?""+t[e]:W.join(String(e)))),t===C?t[e]=r:n?t[e]?t[e]=r:D(t,e,r):(delete t[e],D(t,e,r)))})(Function.prototype,"toString",(function(){return"function"==typeof this&&this[q]||V.call(this)}));var H=function(t){if("function"!=typeof t)throw TypeError(t+" is not a function!");return t},Y=H,$=function(t,e,r){if(Y(t),void 0===e)return t;switch(r){case 1:return function(r){return t.call(e,r)};case 2:return function(r,n){return t.call(e,r,n)};case 3:return function(r,n,o){return t.call(e,r,n,o)}}return function(){return t.apply(e,arguments)}},J=n.exports,z=c.exports,X=j,K=I.exports,Q=$,Z=function(t,e,r){var n,o,i,a,u=t&Z.F,s=t&Z.G,c=t&Z.S,f=t&Z.P,l=t&Z.B,h=s?J:c?J[e]||(J[e]={}):(J[e]||{}).prototype,p=s?z:z[e]||(z[e]={}),y=p.prototype||(p.prototype={});for(n in s&&(r=e),r)i=((o=!u&&h&&void 0!==h[n])?h:r)[n],a=l&&o?Q(i,J):f&&"function"==typeof i?Q(Function.call,i):i,h&&K(h,n,i,t&Z.U),p[n]!=i&&X(p,n,a),f&&y[n]!=i&&(y[n]=i)};J.core=z,Z.F=1,Z.G=2,Z.S=4,Z.P=8,Z.B=16,Z.W=32,Z.U=64,Z.R=128;var tt=Z,et={exports:{}},rt=R("meta"),nt=h,ot=a,it=l.f,at=0,ut=Object.isExtensible||function(){return!0},st=!u((function(){return ut(Object.preventExtensions({}))})),ct=function(t){it(t,rt,{value:{i:"O"+ ++at,w:{}}})},ft=et.exports={KEY:rt,NEED:!1,fastKey:function(t,e){if(!nt(t))return"symbol"==typeof t?t:("string"==typeof t?"S":"P")+t;if(!ot(t,rt)){if(!ut(t))return"F";if(!e)return"E";ct(t)}return t[rt].i},getWeak:function(t,e){if(!ot(t,rt)){if(!ut(t))return!0;if(!e)return!1;ct(t)}return t[rt].w},onFreeze:function(t){return st&&ft.NEED&&ut(t)&&!ot(t,rt)&&ct(t),t}},lt={exports:{}},ht=N.exports("wks"),pt=R,yt=n.exports.Symbol,dt="function"==typeof yt;(lt.exports=function(t){return ht[t]||(ht[t]=dt&&yt[t]||(dt?yt:pt)("Symbol."+t))}).store=ht;var vt=l.f,gt=a,bt=lt.exports("toStringTag"),mt=function(t,e,r){t&&!gt(t=r?t:t.prototype,bt)&&vt(t,bt,{configurable:!0,value:e})},wt={};wt.f=lt.exports;var xt=n.exports,_t=c.exports,St=wt,Et=l.f,At=function(t){var e=_t.Symbol||(_t.Symbol=xt.Symbol||{});"_"==t.charAt(0)||t in e||Et(e,t,{value:St.f(t)})},Ot={}.toString,Tt=function(t){return Ot.call(t).slice(8,-1)},Pt=Tt,jt=Object("z").propertyIsEnumerable(0)?Object:function(t){return"String"==Pt(t)?t.split(""):Object(t)},It=function(t){if(null==t)throw TypeError("Can't call method on  "+t);return t},Lt=jt,Ft=It,Rt=function(t){return Lt(Ft(t))},Nt=Math.ceil,kt=Math.floor,Bt=function(t){return isNaN(t=+t)?0:(t>0?kt:Nt)(t)},Mt=Bt,Ut=Math.min,Ct=function(t){return t>0?Ut(Mt(t),9007199254740991):0},Dt=Bt,Gt=Math.max,qt=Math.min,Vt=function(t,e){return(t=Dt(t))<0?Gt(t+e,0):qt(t,e)},Wt=Rt,Ht=Ct,Yt=Vt,$t=function(t){return function(e,r,n){var o,i=Wt(e),a=Ht(i.length),u=Yt(n,a);if(t&&r!=r){for(;a>u;)if((o=i[u++])!=o)return!0}else for(;a>u;u++)if((t||u in i)&&i[u]===r)return t||u||0;return!t&&-1}},Jt=N.exports("keys"),zt=R,Xt=function(t){return Jt[t]||(Jt[t]=zt(t))},Kt=a,Qt=Rt,Zt=$t(!1),te=Xt("IE_PROTO"),ee=function(t,e){var r,n=Qt(t),o=0,i=[];for(r in n)r!=te&&Kt(n,r)&&i.push(r);for(;e.length>o;)Kt(n,r=e[o++])&&(~Zt(i,r)||i.push(r));return i},re="constructor,hasOwnProperty,isPrototypeOf,propertyIsEnumerable,toLocaleString,toString,valueOf".split(","),ne=ee,oe=re,ie=Object.keys||function(t){return ne(t,oe)},ae={};ae.f=Object.getOwnPropertySymbols;var ue={};ue.f={}.propertyIsEnumerable;var se=ie,ce=ae,fe=ue,le=Tt,he=Array.isArray||function(t){return"Array"==le(t)},pe=It,ye=function(t){return Object(pe(t))},de=l,ve=y,ge=ie,be=s?Object.defineProperties:function(t,e){ve(t);for(var r,n=ge(e),o=n.length,i=0;o>i;)de.f(t,r=n[i++],e[r]);return t},me=n.exports.document,we=me&&me.documentElement,xe=y,_e=be,Se=re,Ee=Xt("IE_PROTO"),Ae=function(){},Oe=function(){var t,e=b("iframe"),r=Se.length;for(e.style.display="none",we.appendChild(e),e.src="javascript:",(t=e.contentWindow.document).open(),t.write("<script>document.F=Object<\/script>"),t.close(),Oe=t.F;r--;)delete Oe.prototype[Se[r]];return Oe()},Te=Object.create||function(t,e){var r;return null!==t?(Ae.prototype=xe(t),r=new Ae,Ae.prototype=null,r[Ee]=t):r=Oe(),void 0===e?r:_e(r,e)},Pe={},je={},Ie=ee,Le=re.concat("length","prototype");je.f=Object.getOwnPropertyNames||function(t){return Ie(t,Le)};var Fe=Rt,Re=je.f,Ne={}.toString,ke="object"==typeof window&&window&&Object.getOwnPropertyNames?Object.getOwnPropertyNames(window):[];Pe.f=function(t){return ke&&"[object Window]"==Ne.call(t)?function(t){try{return Re(t)}catch(t){return ke.slice()}}(t):Re(Fe(t))};var Be={},Me=ue,Ue=O,Ce=Rt,De=x,Ge=a,qe=m,Ve=Object.getOwnPropertyDescriptor;Be.f=s?Ve:function(t,e){if(t=Ce(t),e=De(e,!0),qe)try{return Ve(t,e)}catch(t){}if(Ge(t,e))return Ue(!Me.f.call(t,e),t[e])};var We=n.exports,He=a,Ye=s,$e=tt,Je=I.exports,ze=et.exports.KEY,Xe=u,Ke=N.exports,Qe=mt,Ze=R,tr=lt.exports,er=wt,rr=At,nr=function(t){var e=se(t),r=ce.f;if(r)for(var n,o=r(t),i=fe.f,a=0;o.length>a;)i.call(t,n=o[a++])&&e.push(n);return e},or=he,ir=y,ar=h,ur=ye,sr=Rt,cr=x,fr=O,lr=Te,hr=Pe,pr=Be,yr=ae,dr=l,vr=ie,gr=pr.f,br=dr.f,mr=hr.f,wr=We.Symbol,xr=We.JSON,_r=xr&&xr.stringify,Sr=tr("_hidden"),Er=tr("toPrimitive"),Ar={}.propertyIsEnumerable,Or=Ke("symbol-registry"),Tr=Ke("symbols"),Pr=Ke("op-symbols"),jr=Object.prototype,Ir="function"==typeof wr&&!!yr.f,Lr=We.QObject,Fr=!Lr||!Lr.prototype||!Lr.prototype.findChild,Rr=Ye&&Xe((function(){return 7!=lr(br({},"a",{get:function(){return br(this,"a",{value:7}).a}})).a}))?function(t,e,r){var n=gr(jr,e);n&&delete jr[e],br(t,e,r),n&&t!==jr&&br(jr,e,n)}:br,Nr=function(t){var e=Tr[t]=lr(wr.prototype);return e._k=t,e},kr=Ir&&"symbol"==typeof wr.iterator?function(t){return"symbol"==typeof t}:function(t){return t instanceof wr},Br=function(t,e,r){return t===jr&&Br(Pr,e,r),ir(t),e=cr(e,!0),ir(r),He(Tr,e)?(r.enumerable?(He(t,Sr)&&t[Sr][e]&&(t[Sr][e]=!1),r=lr(r,{enumerable:fr(0,!1)})):(He(t,Sr)||br(t,Sr,fr(1,{})),t[Sr][e]=!0),Rr(t,e,r)):br(t,e,r)},Mr=function(t,e){ir(t);for(var r,n=nr(e=sr(e)),o=0,i=n.length;i>o;)Br(t,r=n[o++],e[r]);return t},Ur=function(t){var e=Ar.call(this,t=cr(t,!0));return!(this===jr&&He(Tr,t)&&!He(Pr,t))&&(!(e||!He(this,t)||!He(Tr,t)||He(this,Sr)&&this[Sr][t])||e)},Cr=function(t,e){if(t=sr(t),e=cr(e,!0),t!==jr||!He(Tr,e)||He(Pr,e)){var r=gr(t,e);return!r||!He(Tr,e)||He(t,Sr)&&t[Sr][e]||(r.enumerable=!0),r}},Dr=function(t){for(var e,r=mr(sr(t)),n=[],o=0;r.length>o;)He(Tr,e=r[o++])||e==Sr||e==ze||n.push(e);return n},Gr=function(t){for(var e,r=t===jr,n=mr(r?Pr:sr(t)),o=[],i=0;n.length>i;)!He(Tr,e=n[i++])||r&&!He(jr,e)||o.push(Tr[e]);return o};Ir||(Je((wr=function(){if(this instanceof wr)throw TypeError("Symbol is not a constructor!");var t=Ze(arguments.length>0?arguments[0]:void 0),e=function(r){this===jr&&e.call(Pr,r),He(this,Sr)&&He(this[Sr],t)&&(this[Sr][t]=!1),Rr(this,t,fr(1,r))};return Ye&&Fr&&Rr(jr,t,{configurable:!0,set:e}),Nr(t)}).prototype,"toString",(function(){return this._k})),pr.f=Cr,dr.f=Br,je.f=hr.f=Dr,ue.f=Ur,yr.f=Gr,Ye&&Je(jr,"propertyIsEnumerable",Ur,!0),er.f=function(t){return Nr(tr(t))}),$e($e.G+$e.W+$e.F*!Ir,{Symbol:wr});for(var qr="hasInstance,isConcatSpreadable,iterator,match,replace,search,species,split,toPrimitive,toStringTag,unscopables".split(","),Vr=0;qr.length>Vr;)tr(qr[Vr++]);for(var Wr=vr(tr.store),Hr=0;Wr.length>Hr;)rr(Wr[Hr++]);$e($e.S+$e.F*!Ir,"Symbol",{for:function(t){return He(Or,t+="")?Or[t]:Or[t]=wr(t)},keyFor:function(t){if(!kr(t))throw TypeError(t+" is not a symbol!");for(var e in Or)if(Or[e]===t)return e},useSetter:function(){Fr=!0},useSimple:function(){Fr=!1}}),$e($e.S+$e.F*!Ir,"Object",{create:function(t,e){return void 0===e?lr(t):Mr(lr(t),e)},defineProperty:Br,defineProperties:Mr,getOwnPropertyDescriptor:Cr,getOwnPropertyNames:Dr,getOwnPropertySymbols:Gr});var Yr=Xe((function(){yr.f(1)}));$e($e.S+$e.F*Yr,"Object",{getOwnPropertySymbols:function(t){return yr.f(ur(t))}}),xr&&$e($e.S+$e.F*(!Ir||Xe((function(){var t=wr();return"[null]"!=_r([t])||"{}"!=_r({a:t})||"{}"!=_r(Object(t))}))),"JSON",{stringify:function(t){for(var e,r,n=[t],o=1;arguments.length>o;)n.push(arguments[o++]);if(r=e=n[1],(ar(e)||void 0!==t)&&!kr(t))return or(e)||(e=function(t,e){if("function"==typeof r&&(e=r.call(this,t,e)),!kr(e))return e}),n[1]=e,_r.apply(xr,n)}}),wr.prototype[Er]||j(wr.prototype,Er,wr.prototype.valueOf),Qe(wr,"Symbol"),Qe(Math,"Math",!0),Qe(We.JSON,"JSON",!0);var $r=Bt,Jr=It,zr=function(t){return function(e,r){var n,o,i=String(Jr(e)),a=$r(r),u=i.length;return a<0||a>=u?t?"":void 0:(n=i.charCodeAt(a))<55296||n>56319||a+1===u||(o=i.charCodeAt(a+1))<56320||o>57343?t?i.charAt(a):n:t?i.slice(a,a+2):o-56320+(n-55296<<10)+65536}},Xr={},Kr=Te,Qr=O,Zr=mt,tn={};j(tn,lt.exports("iterator"),(function(){return this}));var en=a,rn=ye,nn=Xt("IE_PROTO"),on=Object.prototype,an=Object.getPrototypeOf||function(t){return t=rn(t),en(t,nn)?t[nn]:"function"==typeof t.constructor&&t instanceof t.constructor?t.constructor.prototype:t instanceof Object?on:null},un=tt,sn=I.exports,cn=j,fn=Xr,ln=function(t,e,r){t.prototype=Kr(tn,{next:Qr(1,r)}),Zr(t,e+" Iterator")},hn=mt,pn=an,yn=lt.exports("iterator"),dn=!([].keys&&"next"in[].keys()),vn=function(){return this},gn=function(t,e,r,n,o,i,a){ln(r,e,n);var u,s,c,f=function(t){if(!dn&&t in y)return y[t];switch(t){case"keys":case"values":return function(){return new r(this,t)}}return function(){return new r(this,t)}},l=e+" Iterator",h="values"==o,p=!1,y=t.prototype,d=y[yn]||y["@@iterator"]||o&&y[o],v=d||f(o),g=o?h?f("entries"):v:void 0,b="Array"==e&&y.entries||d;if(b&&(c=pn(b.call(new t)))!==Object.prototype&&c.next&&(hn(c,l,!0),"function"!=typeof c[yn]&&cn(c,yn,vn)),h&&d&&"values"!==d.name&&(p=!0,v=function(){return d.call(this)}),(dn||p||!y[yn])&&cn(y,yn,v),fn[e]=v,fn[l]=vn,o)if(u={values:h?v:f("values"),keys:i?v:f("keys"),entries:g},a)for(s in u)s in y||sn(y,s,u[s]);else un(un.P+un.F*(dn||p),e,u);return u},bn=zr(!0);gn(String,"String",(function(t){this._t=String(t),this._i=0}),(function(){var t,e=this._t,r=this._i;return r>=e.length?{value:void 0,done:!0}:(t=bn(e,r),this._i+=t.length,{value:t,done:!1})}));var mn=Tt,wn=lt.exports("toStringTag"),xn="Arguments"==mn(function(){return arguments}()),_n=function(t){var e,r,n;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(r=function(t,e){try{return t[e]}catch(t){}}(e=Object(t),wn))?r:xn?mn(e):"Object"==(n=mn(e))&&"function"==typeof e.callee?"Arguments":n},Sn=_n,En={};En[lt.exports("toStringTag")]="z",En+""!="[object z]"&&I.exports(Object.prototype,"toString",(function(){return"[object "+Sn(this)+"]"}),!0);var An=lt.exports("unscopables"),On=Array.prototype;null==On[An]&&j(On,An,{});var Tn=function(t){On[An][t]=!0},Pn=function(t,e){return{value:e,done:!!t}},jn=Xr,In=Rt,Ln=gn(Array,"Array",(function(t,e){this._t=In(t),this._i=0,this._k=e}),(function(){var t=this._t,e=this._k,r=this._i++;return!t||r>=t.length?(this._t=void 0,Pn(1)):Pn(0,"keys"==e?r:"values"==e?t[r]:[r,t[r]])}),"values");jn.Arguments=jn.Array,Tn("keys"),Tn("values"),Tn("entries");for(var Fn=Ln,Rn=ie,Nn=I.exports,kn=n.exports,Bn=j,Mn=Xr,Un=lt.exports,Cn=Un("iterator"),Dn=Un("toStringTag"),Gn=Mn.Array,qn={CSSRuleList:!0,CSSStyleDeclaration:!1,CSSValueList:!1,ClientRectList:!1,DOMRectList:!1,DOMStringList:!1,DOMTokenList:!0,DataTransferItemList:!1,FileList:!1,HTMLAllCollection:!1,HTMLCollection:!1,HTMLFormElement:!1,HTMLSelectElement:!1,MediaList:!0,MimeTypeArray:!1,NamedNodeMap:!1,NodeList:!0,PaintRequestList:!1,Plugin:!1,PluginArray:!1,SVGLengthList:!1,SVGNumberList:!1,SVGPathSegList:!1,SVGPointList:!1,SVGStringList:!1,SVGTransformList:!1,SourceBufferList:!1,StyleSheetList:!0,TextTrackCueList:!1,TextTrackList:!1,TouchList:!1},Vn=Rn(qn),Wn=0;Wn<Vn.length;Wn++){var Hn,Yn=Vn[Wn],$n=qn[Yn],Jn=kn[Yn],zn=Jn&&Jn.prototype;if(zn&&(zn[Cn]||Bn(zn,Cn,Gn),zn[Dn]||Bn(zn,Dn,Yn),Mn[Yn]=Gn,$n))for(Hn in Fn)zn[Hn]||Nn(zn,Hn,Fn[Hn],!0)}At("asyncIterator");var Xn=tt,Kn=c.exports,Qn=u,Zn=function(t,e){var r=(Kn.Object||{})[t]||Object[t],n={};n[t]=e(r),Xn(Xn.S+Xn.F*Qn((function(){r(1)})),"Object",n)},to=ye,eo=an;Zn("getPrototypeOf",(function(){return function(t){return eo(to(t))}}));var ro=l.f,no=Function.prototype,oo=/^\s*function ([^ (]*)/;"name"in no||s&&ro(no,"name",{configurable:!0,get:function(){try{return(""+this).match(oo)[1]}catch(t){return""}}});var io=function(t,e,r,n){if(!(t instanceof e)||void 0!==n&&n in t)throw TypeError(r+": incorrect invocation!");return t},ao={exports:{}},uo=y,so=function(t,e,r,n){try{return n?e(uo(r)[0],r[1]):e(r)}catch(e){var o=t.return;throw void 0!==o&&uo(o.call(t)),e}},co=Xr,fo=lt.exports("iterator"),lo=Array.prototype,ho=function(t){return void 0!==t&&(co.Array===t||lo[fo]===t)},po=_n,yo=lt.exports("iterator"),vo=Xr,go=c.exports.getIteratorMethod=function(t){if(null!=t)return t[yo]||t["@@iterator"]||vo[po(t)]},bo=$,mo=so,wo=ho,xo=y,_o=Ct,So=go,Eo={},Ao={},Oo=ao.exports=function(t,e,r,n,o){var i,a,u,s,c=o?function(){return t}:So(t),f=bo(r,n,e?2:1),l=0;if("function"!=typeof c)throw TypeError(t+" is not iterable!");if(wo(c)){for(i=_o(t.length);i>l;l++)if((s=e?f(xo(a=t[l])[0],a[1]):f(t[l]))===Eo||s===Ao)return s}else for(u=c.call(t);!(a=u.next()).done;)if((s=mo(u,f,a.value,e))===Eo||s===Ao)return s};Oo.BREAK=Eo,Oo.RETURN=Ao;var To,Po,jo,Io=y,Lo=H,Fo=lt.exports("species"),Ro=function(t,e){var r,n=Io(t).constructor;return void 0===n||null==(r=Io(n)[Fo])?e:Lo(r)},No=$,ko=function(t,e,r){var n=void 0===r;switch(e.length){case 0:return n?t():t.call(r);case 1:return n?t(e[0]):t.call(r,e[0]);case 2:return n?t(e[0],e[1]):t.call(r,e[0],e[1]);case 3:return n?t(e[0],e[1],e[2]):t.call(r,e[0],e[1],e[2]);case 4:return n?t(e[0],e[1],e[2],e[3]):t.call(r,e[0],e[1],e[2],e[3])}return t.apply(r,e)},Bo=we,Mo=b,Uo=n.exports,Co=Uo.process,Do=Uo.setImmediate,Go=Uo.clearImmediate,qo=Uo.MessageChannel,Vo=Uo.Dispatch,Wo=0,Ho={},Yo=function(){var t=+this;if(Ho.hasOwnProperty(t)){var e=Ho[t];delete Ho[t],e()}},$o=function(t){Yo.call(t.data)};Do&&Go||(Do=function(t){for(var e=[],r=1;arguments.length>r;)e.push(arguments[r++]);return Ho[++Wo]=function(){ko("function"==typeof t?t:Function(t),e)},To(Wo),Wo},Go=function(t){delete Ho[t]},"process"==Tt(Co)?To=function(t){Co.nextTick(No(Yo,t,1))}:Vo&&Vo.now?To=function(t){Vo.now(No(Yo,t,1))}:qo?(jo=(Po=new qo).port2,Po.port1.onmessage=$o,To=No(jo.postMessage,jo,1)):Uo.addEventListener&&"function"==typeof postMessage&&!Uo.importScripts?(To=function(t){Uo.postMessage(t+"","*")},Uo.addEventListener("message",$o,!1)):To="onreadystatechange"in Mo("script")?function(t){Bo.appendChild(Mo("script")).onreadystatechange=function(){Bo.removeChild(this),Yo.call(t)}}:function(t){setTimeout(No(Yo,t,1),0)});var Jo={set:Do,clear:Go},zo=n.exports,Xo=Jo.set,Ko=zo.MutationObserver||zo.WebKitMutationObserver,Qo=zo.process,Zo=zo.Promise,ti="process"==Tt(Qo),ei={},ri=H;function ni(t){var e,r;this.promise=new t((function(t,n){if(void 0!==e||void 0!==r)throw TypeError("Bad Promise constructor");e=t,r=n})),this.resolve=ri(e),this.reject=ri(r)}ei.f=function(t){return new ni(t)};var oi=n.exports.navigator,ii=oi&&oi.userAgent||"",ai=y,ui=h,si=ei,ci=I.exports,fi=function(t,e,r){for(var n in e)ci(t,n,e[n],r);return t},li=n.exports,hi=l,pi=s,yi=lt.exports("species"),di=function(t){var e=li[t];pi&&e&&!e[yi]&&hi.f(e,yi,{configurable:!0,get:function(){return this}})},vi=lt.exports("iterator"),gi=!1;try{var bi=[7][vi]();bi.return=function(){gi=!0},Array.from(bi,(function(){throw 2}))}catch(t){}var mi,wi,xi,_i,Si=function(t,e){if(!e&&!gi)return!1;var r=!1;try{var n=[7],o=n[vi]();o.next=function(){return{done:r=!0}},n[vi]=function(){return o},t(n)}catch(t){}return r},Ei=n.exports,Ai=$,Oi=_n,Ti=tt,Pi=h,ji=H,Ii=io,Li=ao.exports,Fi=Ro,Ri=Jo.set,Ni=function(){var t,e,r,n=function(){var n,o;for(ti&&(n=Qo.domain)&&n.exit();t;){o=t.fn,t=t.next;try{o()}catch(n){throw t?r():e=void 0,n}}e=void 0,n&&n.enter()};if(ti)r=function(){Qo.nextTick(n)};else if(!Ko||zo.navigator&&zo.navigator.standalone)if(Zo&&Zo.resolve){var o=Zo.resolve(void 0);r=function(){o.then(n)}}else r=function(){Xo.call(zo,n)};else{var i=!0,a=document.createTextNode("");new Ko(n).observe(a,{characterData:!0}),r=function(){a.data=i=!i}}return function(n){var o={fn:n,next:void 0};e&&(e.next=o),t||(t=o,r()),e=o}}(),ki=ei,Bi=function(t){try{return{e:!1,v:t()}}catch(t){return{e:!0,v:t}}},Mi=ii,Ui=function(t,e){if(ai(t),ui(e)&&e.constructor===t)return e;var r=si.f(t);return(0,r.resolve)(e),r.promise},Ci=Ei.TypeError,Di=Ei.process,Gi=Di&&Di.versions,qi=Gi&&Gi.v8||"",Vi=Ei.Promise,Wi="process"==Oi(Di),Hi=function(){},Yi=wi=ki.f,$i=!!function(){try{var t=Vi.resolve(1),e=(t.constructor={})[lt.exports("species")]=function(t){t(Hi,Hi)};return(Wi||"function"==typeof PromiseRejectionEvent)&&t.then(Hi)instanceof e&&0!==qi.indexOf("6.6")&&-1===Mi.indexOf("Chrome/66")}catch(t){}}(),Ji=function(t){var e;return!(!Pi(t)||"function"!=typeof(e=t.then))&&e},zi=function(t,e){if(!t._n){t._n=!0;var r=t._c;Ni((function(){for(var n=t._v,o=1==t._s,i=0,a=function(e){var r,i,a,u=o?e.ok:e.fail,s=e.resolve,c=e.reject,f=e.domain;try{u?(o||(2==t._h&&Qi(t),t._h=1),!0===u?r=n:(f&&f.enter(),r=u(n),f&&(f.exit(),a=!0)),r===e.promise?c(Ci("Promise-chain cycle")):(i=Ji(r))?i.call(r,s,c):s(r)):c(n)}catch(t){f&&!a&&f.exit(),c(t)}};r.length>i;)a(r[i++]);t._c=[],t._n=!1,e&&!t._h&&Xi(t)}))}},Xi=function(t){Ri.call(Ei,(function(){var e,r,n,o=t._v,i=Ki(t);if(i&&(e=Bi((function(){Wi?Di.emit("unhandledRejection",o,t):(r=Ei.onunhandledrejection)?r({promise:t,reason:o}):(n=Ei.console)&&n.error&&n.error("Unhandled promise rejection",o)})),t._h=Wi||Ki(t)?2:1),t._a=void 0,i&&e.e)throw e.v}))},Ki=function(t){return 1!==t._h&&0===(t._a||t._c).length},Qi=function(t){Ri.call(Ei,(function(){var e;Wi?Di.emit("rejectionHandled",t):(e=Ei.onrejectionhandled)&&e({promise:t,reason:t._v})}))},Zi=function(t){var e=this;e._d||(e._d=!0,(e=e._w||e)._v=t,e._s=2,e._a||(e._a=e._c.slice()),zi(e,!0))},ta=function(t){var e,r=this;if(!r._d){r._d=!0,r=r._w||r;try{if(r===t)throw Ci("Promise can't be resolved itself");(e=Ji(t))?Ni((function(){var n={_w:r,_d:!1};try{e.call(t,Ai(ta,n,1),Ai(Zi,n,1))}catch(t){Zi.call(n,t)}})):(r._v=t,r._s=1,zi(r,!1))}catch(t){Zi.call({_w:r,_d:!1},t)}}};$i||(Vi=function(t){Ii(this,Vi,"Promise","_h"),ji(t),mi.call(this);try{t(Ai(ta,this,1),Ai(Zi,this,1))}catch(t){Zi.call(this,t)}},(mi=function(t){this._c=[],this._a=void 0,this._s=0,this._d=!1,this._v=void 0,this._h=0,this._n=!1}).prototype=fi(Vi.prototype,{then:function(t,e){var r=Yi(Fi(this,Vi));return r.ok="function"!=typeof t||t,r.fail="function"==typeof e&&e,r.domain=Wi?Di.domain:void 0,this._c.push(r),this._a&&this._a.push(r),this._s&&zi(this,!1),r.promise},catch:function(t){return this.then(void 0,t)}}),xi=function(){var t=new mi;this.promise=t,this.resolve=Ai(ta,t,1),this.reject=Ai(Zi,t,1)},ki.f=Yi=function(t){return t===Vi||t===_i?new xi(t):wi(t)}),Ti(Ti.G+Ti.W+Ti.F*!$i,{Promise:Vi}),mt(Vi,"Promise"),di("Promise"),_i=c.exports.Promise,Ti(Ti.S+Ti.F*!$i,"Promise",{reject:function(t){var e=Yi(this);return(0,e.reject)(t),e.promise}}),Ti(Ti.S+Ti.F*!$i,"Promise",{resolve:function(t){return Ui(this,t)}}),Ti(Ti.S+Ti.F*!($i&&Si((function(t){Vi.all(t).catch(Hi)}))),"Promise",{all:function(t){var e=this,r=Yi(e),n=r.resolve,o=r.reject,i=Bi((function(){var r=[],i=0,a=1;Li(t,!1,(function(t){var u=i++,s=!1;r.push(void 0),a++,e.resolve(t).then((function(t){s||(s=!0,r[u]=t,--a||n(r))}),o)})),--a||n(r)}));return i.e&&o(i.v),r.promise},race:function(t){var e=this,r=Yi(e),n=r.reject,o=Bi((function(){Li(t,!1,(function(t){e.resolve(t).then(r.resolve,n)}))}));return o.e&&n(o.v),r.promise}});var ea=y,ra=function(){var t=ea(this),e="";return t.global&&(e+="g"),t.ignoreCase&&(e+="i"),t.multiline&&(e+="m"),t.unicode&&(e+="u"),t.sticky&&(e+="y"),e};s&&"g"!=/./g.flags&&l.f(RegExp.prototype,"flags",{configurable:!0,get:ra});var na=y,oa=ra,ia=s,aa=/./.toString,ua=function(t){I.exports(RegExp.prototype,"toString",t,!0)};u((function(){return"/a/b"!=aa.call({source:"a",flags:"b"})}))?ua((function(){var t=na(this);return"/".concat(t.source,"/","flags"in t?t.flags:!ia&&t instanceof RegExp?oa.call(t):void 0)})):"toString"!=aa.name&&ua((function(){return aa.call(this)}));var sa=tt,ca=we,fa=Tt,la=Vt,ha=Ct,pa=[].slice;sa(sa.P+sa.F*u((function(){ca&&pa.call(ca)})),"Array",{slice:function(t,e){var r=ha(this.length),n=fa(this);if(e=void 0===e?r:e,"Array"==n)return pa.call(this,t,e);for(var o=la(t,r),i=la(e,r),a=ha(i-o),u=new Array(a),s=0;s<a;s++)u[s]="String"==n?this.charAt(o+s):this[o+s];return u}});var ya={exports:{}};!function(t){var e=function(t){var e,n=Object.prototype,o=n.hasOwnProperty,i="function"==typeof Symbol?Symbol:{},a=i.iterator||"@@iterator",u=i.asyncIterator||"@@asyncIterator",s=i.toStringTag||"@@toStringTag";function c(t,e,r){return Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}),t[e]}try{c({},"")}catch(t){c=function(t,e,r){return t[e]=r}}function f(t,e,r,n){var o=e&&e.prototype instanceof g?e:g,i=Object.create(o.prototype),a=new j(n||[]);return i._invoke=function(t,e,r){var n=h;return function(o,i){if(n===y)throw new Error("Generator is already running");if(n===d){if("throw"===o)throw i;return L()}for(r.method=o,r.arg=i;;){var a=r.delegate;if(a){var u=O(a,r);if(u){if(u===v)continue;return u}}if("next"===r.method)r.sent=r._sent=r.arg;else if("throw"===r.method){if(n===h)throw n=d,r.arg;r.dispatchException(r.arg)}else"return"===r.method&&r.abrupt("return",r.arg);n=y;var s=l(t,e,r);if("normal"===s.type){if(n=r.done?d:p,s.arg===v)continue;return{value:s.arg,done:r.done}}"throw"===s.type&&(n=d,r.method="throw",r.arg=s.arg)}}}(t,r,a),i}function l(t,e,r){try{return{type:"normal",arg:t.call(e,r)}}catch(t){return{type:"throw",arg:t}}}t.wrap=f;var h="suspendedStart",p="suspendedYield",y="executing",d="completed",v={};function g(){}function b(){}function m(){}var w={};w[a]=function(){return this};var x=Object.getPrototypeOf,_=x&&x(x(I([])));_&&_!==n&&o.call(_,a)&&(w=_);var S=m.prototype=g.prototype=Object.create(w);function E(t){["next","throw","return"].forEach((function(e){c(t,e,(function(t){return this._invoke(e,t)}))}))}function A(t,e){function n(i,a,u,s){var c=l(t[i],t,a);if("throw"!==c.type){var f=c.arg,h=f.value;return h&&"object"===r(h)&&o.call(h,"__await")?e.resolve(h.__await).then((function(t){n("next",t,u,s)}),(function(t){n("throw",t,u,s)})):e.resolve(h).then((function(t){f.value=t,u(f)}),(function(t){return n("throw",t,u,s)}))}s(c.arg)}var i;this._invoke=function(t,r){function o(){return new e((function(e,o){n(t,r,e,o)}))}return i=i?i.then(o,o):o()}}function O(t,r){var n=t.iterator[r.method];if(n===e){if(r.delegate=null,"throw"===r.method){if(t.iterator.return&&(r.method="return",r.arg=e,O(t,r),"throw"===r.method))return v;r.method="throw",r.arg=new TypeError("The iterator does not provide a 'throw' method")}return v}var o=l(n,t.iterator,r.arg);if("throw"===o.type)return r.method="throw",r.arg=o.arg,r.delegate=null,v;var i=o.arg;return i?i.done?(r[t.resultName]=i.value,r.next=t.nextLoc,"return"!==r.method&&(r.method="next",r.arg=e),r.delegate=null,v):i:(r.method="throw",r.arg=new TypeError("iterator result is not an object"),r.delegate=null,v)}function T(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function P(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function j(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(T,this),this.reset(!0)}function I(t){if(t){var r=t[a];if(r)return r.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var n=-1,i=function r(){for(;++n<t.length;)if(o.call(t,n))return r.value=t[n],r.done=!1,r;return r.value=e,r.done=!0,r};return i.next=i}}return{next:L}}function L(){return{value:e,done:!0}}return b.prototype=S.constructor=m,m.constructor=b,b.displayName=c(m,s,"GeneratorFunction"),t.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===b||"GeneratorFunction"===(e.displayName||e.name))},t.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,m):(t.__proto__=m,c(t,s,"GeneratorFunction")),t.prototype=Object.create(S),t},t.awrap=function(t){return{__await:t}},E(A.prototype),A.prototype[u]=function(){return this},t.AsyncIterator=A,t.async=function(e,r,n,o,i){void 0===i&&(i=Promise);var a=new A(f(e,r,n,o),i);return t.isGeneratorFunction(r)?a:a.next().then((function(t){return t.done?t.value:a.next()}))},E(S),c(S,s,"Generator"),S[a]=function(){return this},S.toString=function(){return"[object Generator]"},t.keys=function(t){var e=[];for(var r in t)e.push(r);return e.reverse(),function r(){for(;e.length;){var n=e.pop();if(n in t)return r.value=n,r.done=!1,r}return r.done=!0,r}},t.values=I,j.prototype={constructor:j,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=e,this.done=!1,this.delegate=null,this.method="next",this.arg=e,this.tryEntries.forEach(P),!t)for(var r in this)"t"===r.charAt(0)&&o.call(this,r)&&!isNaN(+r.slice(1))&&(this[r]=e)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var r=this;function n(n,o){return u.type="throw",u.arg=t,r.next=n,o&&(r.method="next",r.arg=e),!!o}for(var i=this.tryEntries.length-1;i>=0;--i){var a=this.tryEntries[i],u=a.completion;if("root"===a.tryLoc)return n("end");if(a.tryLoc<=this.prev){var s=o.call(a,"catchLoc"),c=o.call(a,"finallyLoc");if(s&&c){if(this.prev<a.catchLoc)return n(a.catchLoc,!0);if(this.prev<a.finallyLoc)return n(a.finallyLoc)}else if(s){if(this.prev<a.catchLoc)return n(a.catchLoc,!0)}else{if(!c)throw new Error("try statement without catch or finally");if(this.prev<a.finallyLoc)return n(a.finallyLoc)}}}},abrupt:function(t,e){for(var r=this.tryEntries.length-1;r>=0;--r){var n=this.tryEntries[r];if(n.tryLoc<=this.prev&&o.call(n,"finallyLoc")&&this.prev<n.finallyLoc){var i=n;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=e&&e<=i.finallyLoc&&(i=null);var a=i?i.completion:{};return a.type=t,a.arg=e,i?(this.method="next",this.next=i.finallyLoc,v):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),v},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.finallyLoc===t)return this.complete(r.completion,r.afterLoc),P(r),v}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.tryLoc===t){var n=r.completion;if("throw"===n.type){var o=n.arg;P(r)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,r,n){return this.delegate={iterator:I(t),resultName:r,nextLoc:n},"next"===this.method&&(this.arg=e),v}},t}(t.exports);try{regeneratorRuntime=e}catch(t){Function("r","regeneratorRuntime = r")(e)}}(ya);var da=ya.exports;tt(tt.S,"Number",{MAX_SAFE_INTEGER:9007199254740991});var va=h,ga=y,ba=function(t,e){if(ga(t),!va(e)&&null!==e)throw TypeError(e+": can't set as prototype!")},ma={set:Object.setPrototypeOf||("__proto__"in{}?function(t,e,r){try{(r=$(Function.call,Be.f(Object.prototype,"__proto__").set,2))(t,[]),e=!(t instanceof Array)}catch(t){e=!0}return function(t,n){return ba(t,n),e?t.__proto__=n:r(t,n),t}}({},!1):void 0),check:ba},wa=h,xa=ma.set,_a=tt,Sa=It,Ea=u,Aa="\t\n\v\f\r   ᠎             　\u2028\u2029\ufeff",Oa="["+Aa+"]",Ta=RegExp("^"+Oa+Oa+"*"),Pa=RegExp(Oa+Oa+"*$"),ja=function(t,e,r){var n={},o=Ea((function(){return!!Aa[t]()||"​"!="​"[t]()})),i=n[t]=o?e(Ia):Aa[t];r&&(n[r]=i),_a(_a.P+_a.F*o,"String",n)},Ia=ja.trim=function(t,e){return t=String(Sa(t)),1&e&&(t=t.replace(Ta,"")),2&e&&(t=t.replace(Pa,"")),t},La=ja,Fa=n.exports,Ra=a,Na=Tt,ka=function(t,e,r){var n,o=e.constructor;return o!==r&&"function"==typeof o&&(n=o.prototype)!==r.prototype&&wa(n)&&xa&&xa(t,n),t},Ba=x,Ma=u,Ua=je.f,Ca=Be.f,Da=l.f,Ga=La.trim,qa=Fa.Number,Va=qa,Wa=qa.prototype,Ha="Number"==Na(Te(Wa)),Ya="trim"in String.prototype,$a=function(t){var e=Ba(t,!1);if("string"==typeof e&&e.length>2){var r,n,o,i=(e=Ya?e.trim():Ga(e,3)).charCodeAt(0);if(43===i||45===i){if(88===(r=e.charCodeAt(2))||120===r)return NaN}else if(48===i){switch(e.charCodeAt(1)){case 66:case 98:n=2,o=49;break;case 79:case 111:n=8,o=55;break;default:return+e}for(var a,u=e.slice(2),s=0,c=u.length;s<c;s++)if((a=u.charCodeAt(s))<48||a>o)return NaN;return parseInt(u,n)}}return+e};if(!qa(" 0o1")||!qa("0b1")||qa("+0x1")){qa=function(t){var e=arguments.length<1?0:t,r=this;return r instanceof qa&&(Ha?Ma((function(){Wa.valueOf.call(r)})):"Number"!=Na(r))?ka(new Va($a(e)),r,qa):$a(e)};for(var Ja,za=s?Ua(Va):"MAX_VALUE,MIN_VALUE,NaN,NEGATIVE_INFINITY,POSITIVE_INFINITY,EPSILON,isFinite,isInteger,isNaN,isSafeInteger,MAX_SAFE_INTEGER,MIN_SAFE_INTEGER,parseFloat,parseInt,isInteger".split(","),Xa=0;za.length>Xa;Xa++)Ra(Va,Ja=za[Xa])&&!Ra(qa,Ja)&&Da(qa,Ja,Ca(Va,Ja));qa.prototype=Wa,Wa.constructor=qa,I.exports(Fa,"Number",qa)}var Ka,Qa,Za=h,tu=Tt,eu=lt.exports("match"),ru=zr(!0),nu=function(t,e,r){return e+(r?ru(t,e).length:1)},ou=_n,iu=RegExp.prototype.exec,au=function(t,e){var r=t.exec;if("function"==typeof r){var n=r.call(t,e);if("object"!=typeof n)throw new TypeError("RegExp exec method returned something other than an Object or null");return n}if("RegExp"!==ou(t))throw new TypeError("RegExp#exec called on incompatible receiver");return iu.call(t,e)},uu=ra,su=RegExp.prototype.exec,cu=String.prototype.replace,fu=su,lu=(Ka=/a/,Qa=/b*/g,su.call(Ka,"a"),su.call(Qa,"a"),0!==Ka.lastIndex||0!==Qa.lastIndex),hu=void 0!==/()??/.exec("")[1];(lu||hu)&&(fu=function(t){var e,r,n,o,i=this;return hu&&(r=new RegExp("^"+i.source+"$(?!\\s)",uu.call(i))),lu&&(e=i.lastIndex),n=su.call(i,t),lu&&n&&(i.lastIndex=i.global?n.index+n[0].length:e),hu&&n&&n.length>1&&cu.call(n[0],r,(function(){for(o=1;o<arguments.length-2;o++)void 0===arguments[o]&&(n[o]=void 0)})),n});var pu=fu;tt({target:"RegExp",proto:!0,forced:pu!==/./.exec},{exec:pu});var yu=I.exports,du=j,vu=u,gu=It,bu=lt.exports,mu=pu,wu=bu("species"),xu=!vu((function(){var t=/./;return t.exec=function(){var t=[];return t.groups={a:"7"},t},"7"!=="".replace(t,"$<a>")})),_u=function(){var t=/(?:)/,e=t.exec;t.exec=function(){return e.apply(this,arguments)};var r="ab".split(t);return 2===r.length&&"a"===r[0]&&"b"===r[1]}(),Su=function(t,e,r){var n=bu(t),o=!vu((function(){var e={};return e[n]=function(){return 7},7!=""[t](e)})),i=o?!vu((function(){var e=!1,r=/a/;return r.exec=function(){return e=!0,null},"split"===t&&(r.constructor={},r.constructor[wu]=function(){return r}),r[n](""),!e})):void 0;if(!o||!i||"replace"===t&&!xu||"split"===t&&!_u){var a=/./[n],u=r(gu,n,""[t],(function(t,e,r,n,i){return e.exec===mu?o&&!i?{done:!0,value:a.call(e,r,n)}:{done:!0,value:t.call(r,e,n)}:{done:!1}})),s=u[0],c=u[1];yu(String.prototype,t,s),du(RegExp.prototype,n,2==e?function(t,e){return c.call(t,this,e)}:function(t){return c.call(t,this)})}},Eu=function(t){var e;return Za(t)&&(void 0!==(e=t[eu])?!!e:"RegExp"==tu(t))},Au=y,Ou=Ro,Tu=nu,Pu=Ct,ju=au,Iu=pu,Lu=u,Fu=Math.min,Ru=[].push,Nu=!Lu((function(){RegExp(4294967295,"y")}));Su("split",2,(function(t,e,r,n){var o;return o="c"=="abbc".split(/(b)*/)[1]||4!="test".split(/(?:)/,-1).length||2!="ab".split(/(?:ab)*/).length||4!=".".split(/(.?)(.?)/).length||".".split(/()()/).length>1||"".split(/.?/).length?function(t,e){var n=String(this);if(void 0===t&&0===e)return[];if(!Eu(t))return r.call(n,t,e);for(var o,i,a,u=[],s=(t.ignoreCase?"i":"")+(t.multiline?"m":"")+(t.unicode?"u":"")+(t.sticky?"y":""),c=0,f=void 0===e?4294967295:e>>>0,l=new RegExp(t.source,s+"g");(o=Iu.call(l,n))&&!((i=l.lastIndex)>c&&(u.push(n.slice(c,o.index)),o.length>1&&o.index<n.length&&Ru.apply(u,o.slice(1)),a=o[0].length,c=i,u.length>=f));)l.lastIndex===o.index&&l.lastIndex++;return c===n.length?!a&&l.test("")||u.push(""):u.push(n.slice(c)),u.length>f?u.slice(0,f):u}:"0".split(void 0,0).length?function(t,e){return void 0===t&&0===e?[]:r.call(this,t,e)}:r,[function(r,n){var i=t(this),a=null==r?void 0:r[e];return void 0!==a?a.call(r,i,n):o.call(String(i),r,n)},function(t,e){var i=n(o,t,this,e,o!==r);if(i.done)return i.value;var a=Au(t),u=String(this),s=Ou(a,RegExp),c=a.unicode,f=(a.ignoreCase?"i":"")+(a.multiline?"m":"")+(a.unicode?"u":"")+(Nu?"y":"g"),l=new s(Nu?a:"^(?:"+a.source+")",f),h=void 0===e?4294967295:e>>>0;if(0===h)return[];if(0===u.length)return null===ju(l,u)?[u]:[];for(var p=0,y=0,d=[];y<u.length;){l.lastIndex=Nu?y:0;var v,g=ju(l,Nu?u:u.slice(y));if(null===g||(v=Fu(Pu(l.lastIndex+(Nu?0:y)),u.length))===p)y=Tu(u,y,c);else{if(d.push(u.slice(p,y)),d.length===h)return d;for(var b=1;b<=g.length-1;b++)if(d.push(g[b]),d.length===h)return d;y=p=v}}return d.push(u.slice(p)),d}]}));var ku=l,Bu=O,Mu=$,Uu=tt,Cu=ye,Du=so,Gu=ho,qu=Ct,Vu=function(t,e,r){e in t?ku.f(t,e,Bu(0,r)):t[e]=r},Wu=go;Uu(Uu.S+Uu.F*!Si((function(t){Array.from(t)})),"Array",{from:function(t){var e,r,n,o,i=Cu(t),a="function"==typeof this?this:Array,u=arguments.length,s=u>1?arguments[1]:void 0,c=void 0!==s,f=0,l=Wu(i);if(c&&(s=Mu(s,u>2?arguments[2]:void 0,2)),null==l||a==Array&&Gu(l))for(r=new a(e=qu(i.length));e>f;f++)Vu(r,f,c?s(i[f],f):i[f]);else for(o=l.call(i),r=new a;!(n=o.next()).done;f++)Vu(r,f,c?Du(o,s,[n.value,f],!0):n.value);return r.length=f,r}});var Hu,Yu,$u=h,Ju=he,zu=lt.exports("species"),Xu=function(t){var e;return Ju(t)&&("function"!=typeof(e=t.constructor)||e!==Array&&!Ju(e.prototype)||(e=void 0),$u(e)&&null===(e=e[zu])&&(e=void 0)),void 0===e?Array:e},Ku=$,Qu=jt,Zu=ye,ts=Ct,es=function(t,e){return new(Xu(t))(e)},rs=function(t,e){var r=1==t,n=2==t,o=3==t,i=4==t,a=6==t,u=5==t||a,s=e||es;return function(e,c,f){for(var l,h,p=Zu(e),y=Qu(p),d=Ku(c,f,3),v=ts(y.length),g=0,b=r?s(e,v):n?s(e,0):void 0;v>g;g++)if((u||g in y)&&(h=d(l=y[g],g,p),t))if(r)b[g]=h;else if(h)switch(t){case 3:return!0;case 5:return l;case 6:return g;case 2:b.push(l)}else if(i)return!1;return a?-1:o||i?i:b}},ns=u,os=tt,is=rs(1);os(os.P+os.F*(Yu=!0,!((Hu=[].map)&&ns((function(){Yu?Hu.call(null,(function(){}),1):Hu.call(null)})))),"Array",{map:function(t){return is(this,t,arguments[1])}}),Zn("getOwnPropertyNames",(function(){return Pe.f}));for(var as,us={exports:{}},ss=n.exports,cs=j,fs=R,ls=fs("typed_array"),hs=fs("view"),ps=!(!ss.ArrayBuffer||!ss.DataView),ys=ps,ds=0,vs="Int8Array,Uint8Array,Uint8ClampedArray,Int16Array,Uint16Array,Int32Array,Uint32Array,Float32Array,Float64Array".split(",");ds<9;)(as=ss[vs[ds++]])?(cs(as.prototype,ls,!0),cs(as.prototype,hs,!0)):ys=!1;var gs={ABV:ps,CONSTR:ys,TYPED:ls,VIEW:hs},bs={},ms=Bt,ws=Ct,xs=function(t){if(void 0===t)return 0;var e=ms(t),r=ws(e);if(e!==r)throw RangeError("Wrong length!");return r},_s=ye,Ss=Vt,Es=Ct,As=function(t){for(var e=_s(this),r=Es(e.length),n=arguments.length,o=Ss(n>1?arguments[1]:void 0,r),i=n>2?arguments[2]:void 0,a=void 0===i?r:Ss(i,r);a>o;)e[o++]=t;return e};!function(t){var e=n.exports,r=s,o=gs,i=j,a=fi,c=u,f=io,h=Bt,p=Ct,y=xs,d=je.f,v=l.f,g=As,b=mt,m="ArrayBuffer",w="DataView",x="Wrong index!",_=e.ArrayBuffer,S=e.DataView,E=e.Math,A=e.RangeError,O=e.Infinity,T=_,P=E.abs,I=E.pow,L=E.floor,F=E.log,R=E.LN2,N="buffer",k="byteLength",B="byteOffset",M=r?"_b":N,U=r?"_l":k,C=r?"_o":B;function D(t,e,r){var n,o,i,a=new Array(r),u=8*r-e-1,s=(1<<u)-1,c=s>>1,f=23===e?I(2,-24)-I(2,-77):0,l=0,h=t<0||0===t&&1/t<0?1:0;for((t=P(t))!=t||t===O?(o=t!=t?1:0,n=s):(n=L(F(t)/R),t*(i=I(2,-n))<1&&(n--,i*=2),(t+=n+c>=1?f/i:f*I(2,1-c))*i>=2&&(n++,i/=2),n+c>=s?(o=0,n=s):n+c>=1?(o=(t*i-1)*I(2,e),n+=c):(o=t*I(2,c-1)*I(2,e),n=0));e>=8;a[l++]=255&o,o/=256,e-=8);for(n=n<<e|o,u+=e;u>0;a[l++]=255&n,n/=256,u-=8);return a[--l]|=128*h,a}function G(t,e,r){var n,o=8*r-e-1,i=(1<<o)-1,a=i>>1,u=o-7,s=r-1,c=t[s--],f=127&c;for(c>>=7;u>0;f=256*f+t[s],s--,u-=8);for(n=f&(1<<-u)-1,f>>=-u,u+=e;u>0;n=256*n+t[s],s--,u-=8);if(0===f)f=1-a;else{if(f===i)return n?NaN:c?-O:O;n+=I(2,e),f-=a}return(c?-1:1)*n*I(2,f-e)}function q(t){return t[3]<<24|t[2]<<16|t[1]<<8|t[0]}function V(t){return[255&t]}function W(t){return[255&t,t>>8&255]}function H(t){return[255&t,t>>8&255,t>>16&255,t>>24&255]}function Y(t){return D(t,52,8)}function $(t){return D(t,23,4)}function J(t,e,r){v(t.prototype,e,{get:function(){return this[r]}})}function z(t,e,r,n){var o=y(+r);if(o+e>t[U])throw A(x);var i=t[M]._b,a=o+t[C],u=i.slice(a,a+e);return n?u:u.reverse()}function X(t,e,r,n,o,i){var a=y(+r);if(a+e>t[U])throw A(x);for(var u=t[M]._b,s=a+t[C],c=n(+o),f=0;f<e;f++)u[s+f]=c[i?f:e-f-1]}if(o.ABV){if(!c((function(){_(1)}))||!c((function(){new _(-1)}))||c((function(){return new _,new _(1.5),new _(NaN),_.name!=m}))){for(var K,Q=(_=function(t){return f(this,_),new T(y(t))}).prototype=T.prototype,Z=d(T),tt=0;Z.length>tt;)(K=Z[tt++])in _||i(_,K,T[K]);Q.constructor=_}var et=new S(new _(2)),rt=S.prototype.setInt8;et.setInt8(0,2147483648),et.setInt8(1,2147483649),!et.getInt8(0)&&et.getInt8(1)||a(S.prototype,{setInt8:function(t,e){rt.call(this,t,e<<24>>24)},setUint8:function(t,e){rt.call(this,t,e<<24>>24)}},!0)}else _=function(t){f(this,_,m);var e=y(t);this._b=g.call(new Array(e),0),this[U]=e},S=function(t,e,r){f(this,S,w),f(t,_,w);var n=t[U],o=h(e);if(o<0||o>n)throw A("Wrong offset!");if(o+(r=void 0===r?n-o:p(r))>n)throw A("Wrong length!");this[M]=t,this[C]=o,this[U]=r},r&&(J(_,k,"_l"),J(S,N,"_b"),J(S,k,"_l"),J(S,B,"_o")),a(S.prototype,{getInt8:function(t){return z(this,1,t)[0]<<24>>24},getUint8:function(t){return z(this,1,t)[0]},getInt16:function(t){var e=z(this,2,t,arguments[1]);return(e[1]<<8|e[0])<<16>>16},getUint16:function(t){var e=z(this,2,t,arguments[1]);return e[1]<<8|e[0]},getInt32:function(t){return q(z(this,4,t,arguments[1]))},getUint32:function(t){return q(z(this,4,t,arguments[1]))>>>0},getFloat32:function(t){return G(z(this,4,t,arguments[1]),23,4)},getFloat64:function(t){return G(z(this,8,t,arguments[1]),52,8)},setInt8:function(t,e){X(this,1,t,V,e)},setUint8:function(t,e){X(this,1,t,V,e)},setInt16:function(t,e){X(this,2,t,W,e,arguments[2])},setUint16:function(t,e){X(this,2,t,W,e,arguments[2])},setInt32:function(t,e){X(this,4,t,H,e,arguments[2])},setUint32:function(t,e){X(this,4,t,H,e,arguments[2])},setFloat32:function(t,e){X(this,4,t,$,e,arguments[2])},setFloat64:function(t,e){X(this,8,t,Y,e,arguments[2])}});b(_,m),b(S,w),i(S.prototype,o.VIEW,!0),t.ArrayBuffer=_,t.DataView=S}(bs);var Os=ye,Ts=Vt,Ps=Ct,js=[].copyWithin||function(t,e){var r=Os(this),n=Ps(r.length),o=Ts(t,n),i=Ts(e,n),a=arguments.length>2?arguments[2]:void 0,u=Math.min((void 0===a?n:Ts(a,n))-i,n-o),s=1;for(i<o&&o<i+u&&(s=-1,i+=u-1,o+=u-1);u-- >0;)i in r?r[o]=r[i]:delete r[o],o+=s,i+=s;return r};if(s){var Is=n.exports,Ls=u,Fs=tt,Rs=gs,Ns=bs,ks=$,Bs=io,Ms=O,Us=j,Cs=fi,Ds=Bt,Gs=Ct,qs=xs,Vs=Vt,Ws=x,Hs=a,Ys=_n,$s=h,Js=ye,zs=ho,Xs=Te,Ks=an,Qs=je.f,Zs=go,tc=R,ec=lt.exports,rc=rs,nc=$t,oc=Ro,ic=Ln,ac=Xr,uc=Si,sc=di,cc=As,fc=js,lc=l,hc=Be,pc=lc.f,yc=hc.f,dc=Is.RangeError,vc=Is.TypeError,gc=Is.Uint8Array,bc=Array.prototype,mc=Ns.ArrayBuffer,wc=Ns.DataView,xc=rc(0),_c=rc(2),Sc=rc(3),Ec=rc(4),Ac=rc(5),Oc=rc(6),Tc=nc(!0),Pc=nc(!1),jc=ic.values,Ic=ic.keys,Lc=ic.entries,Fc=bc.lastIndexOf,Rc=bc.reduce,Nc=bc.reduceRight,kc=bc.join,Bc=bc.sort,Mc=bc.slice,Uc=bc.toString,Cc=bc.toLocaleString,Dc=ec("iterator"),Gc=ec("toStringTag"),qc=tc("typed_constructor"),Vc=tc("def_constructor"),Wc=Rs.CONSTR,Hc=Rs.TYPED,Yc=Rs.VIEW,$c=rc(1,(function(t,e){return Qc(oc(t,t[Vc]),e)})),Jc=Ls((function(){return 1===new gc(new Uint16Array([1]).buffer)[0]})),zc=!!gc&&!!gc.prototype.set&&Ls((function(){new gc(1).set({})})),Xc=function(t,e){var r=Ds(t);if(r<0||r%e)throw dc("Wrong offset!");return r},Kc=function(t){if($s(t)&&Hc in t)return t;throw vc(t+" is not a typed array!")},Qc=function(t,e){if(!$s(t)||!(qc in t))throw vc("It is not a typed array constructor!");return new t(e)},Zc=function(t,e){return tf(oc(t,t[Vc]),e)},tf=function(t,e){for(var r=0,n=e.length,o=Qc(t,n);n>r;)o[r]=e[r++];return o},ef=function(t,e,r){pc(t,e,{get:function(){return this._d[r]}})},rf=function(t){var e,r,n,o,i,a,u=Js(t),s=arguments.length,c=s>1?arguments[1]:void 0,f=void 0!==c,l=Zs(u);if(null!=l&&!zs(l)){for(a=l.call(u),n=[],e=0;!(i=a.next()).done;e++)n.push(i.value);u=n}for(f&&s>2&&(c=ks(c,arguments[2],2)),e=0,r=Gs(u.length),o=Qc(this,r);r>e;e++)o[e]=f?c(u[e],e):u[e];return o},nf=function(){for(var t=0,e=arguments.length,r=Qc(this,e);e>t;)r[t]=arguments[t++];return r},of=!!gc&&Ls((function(){Cc.call(new gc(1))})),af=function(){return Cc.apply(of?Mc.call(Kc(this)):Kc(this),arguments)},uf={copyWithin:function(t,e){return fc.call(Kc(this),t,e,arguments.length>2?arguments[2]:void 0)},every:function(t){return Ec(Kc(this),t,arguments.length>1?arguments[1]:void 0)},fill:function(t){return cc.apply(Kc(this),arguments)},filter:function(t){return Zc(this,_c(Kc(this),t,arguments.length>1?arguments[1]:void 0))},find:function(t){return Ac(Kc(this),t,arguments.length>1?arguments[1]:void 0)},findIndex:function(t){return Oc(Kc(this),t,arguments.length>1?arguments[1]:void 0)},forEach:function(t){xc(Kc(this),t,arguments.length>1?arguments[1]:void 0)},indexOf:function(t){return Pc(Kc(this),t,arguments.length>1?arguments[1]:void 0)},includes:function(t){return Tc(Kc(this),t,arguments.length>1?arguments[1]:void 0)},join:function(t){return kc.apply(Kc(this),arguments)},lastIndexOf:function(t){return Fc.apply(Kc(this),arguments)},map:function(t){return $c(Kc(this),t,arguments.length>1?arguments[1]:void 0)},reduce:function(t){return Rc.apply(Kc(this),arguments)},reduceRight:function(t){return Nc.apply(Kc(this),arguments)},reverse:function(){for(var t,e=this,r=Kc(e).length,n=Math.floor(r/2),o=0;o<n;)t=e[o],e[o++]=e[--r],e[r]=t;return e},some:function(t){return Sc(Kc(this),t,arguments.length>1?arguments[1]:void 0)},sort:function(t){return Bc.call(Kc(this),t)},subarray:function(t,e){var r=Kc(this),n=r.length,o=Vs(t,n);return new(oc(r,r[Vc]))(r.buffer,r.byteOffset+o*r.BYTES_PER_ELEMENT,Gs((void 0===e?n:Vs(e,n))-o))}},sf=function(t,e){return Zc(this,Mc.call(Kc(this),t,e))},cf=function(t){Kc(this);var e=Xc(arguments[1],1),r=this.length,n=Js(t),o=Gs(n.length),i=0;if(o+e>r)throw dc("Wrong length!");for(;i<o;)this[e+i]=n[i++]},ff={entries:function(){return Lc.call(Kc(this))},keys:function(){return Ic.call(Kc(this))},values:function(){return jc.call(Kc(this))}},lf=function(t,e){return $s(t)&&t[Hc]&&"symbol"!=typeof e&&e in t&&String(+e)==String(e)},hf=function(t,e){return lf(t,e=Ws(e,!0))?Ms(2,t[e]):yc(t,e)},pf=function(t,e,r){return!(lf(t,e=Ws(e,!0))&&$s(r)&&Hs(r,"value"))||Hs(r,"get")||Hs(r,"set")||r.configurable||Hs(r,"writable")&&!r.writable||Hs(r,"enumerable")&&!r.enumerable?pc(t,e,r):(t[e]=r.value,t)};Wc||(hc.f=hf,lc.f=pf),Fs(Fs.S+Fs.F*!Wc,"Object",{getOwnPropertyDescriptor:hf,defineProperty:pf}),Ls((function(){Uc.call({})}))&&(Uc=Cc=function(){return kc.call(this)});var yf=Cs({},uf);Cs(yf,ff),Us(yf,Dc,ff.values),Cs(yf,{slice:sf,set:cf,constructor:function(){},toString:Uc,toLocaleString:af}),ef(yf,"buffer","b"),ef(yf,"byteOffset","o"),ef(yf,"byteLength","l"),ef(yf,"length","e"),pc(yf,Gc,{get:function(){return this[Hc]}}),us.exports=function(t,e,r,n){var o=t+((n=!!n)?"Clamped":"")+"Array",i="get"+t,a="set"+t,u=Is[o],s=u||{},c=u&&Ks(u),f=!u||!Rs.ABV,l={},h=u&&u.prototype,p=function(t,r){pc(t,r,{get:function(){return function(t,r){var n=t._d;return n.v[i](r*e+n.o,Jc)}(this,r)},set:function(t){return function(t,r,o){var i=t._d;n&&(o=(o=Math.round(o))<0?0:o>255?255:255&o),i.v[a](r*e+i.o,o,Jc)}(this,r,t)},enumerable:!0})};f?(u=r((function(t,r,n,i){Bs(t,u,o,"_d");var a,s,c,f,l=0,h=0;if($s(r)){if(!(r instanceof mc||"ArrayBuffer"==(f=Ys(r))||"SharedArrayBuffer"==f))return Hc in r?tf(u,r):rf.call(u,r);a=r,h=Xc(n,e);var y=r.byteLength;if(void 0===i){if(y%e)throw dc("Wrong length!");if((s=y-h)<0)throw dc("Wrong length!")}else if((s=Gs(i)*e)+h>y)throw dc("Wrong length!");c=s/e}else c=qs(r),a=new mc(s=c*e);for(Us(t,"_d",{b:a,o:h,l:s,e:c,v:new wc(a)});l<c;)p(t,l++)})),h=u.prototype=Xs(yf),Us(h,"constructor",u)):Ls((function(){u(1)}))&&Ls((function(){new u(-1)}))&&uc((function(t){new u,new u(null),new u(1.5),new u(t)}),!0)||(u=r((function(t,r,n,i){var a;return Bs(t,u,o),$s(r)?r instanceof mc||"ArrayBuffer"==(a=Ys(r))||"SharedArrayBuffer"==a?void 0!==i?new s(r,Xc(n,e),i):void 0!==n?new s(r,Xc(n,e)):new s(r):Hc in r?tf(u,r):rf.call(u,r):new s(qs(r))})),xc(c!==Function.prototype?Qs(s).concat(Qs(c)):Qs(s),(function(t){t in u||Us(u,t,s[t])})),u.prototype=h,h.constructor=u);var y=h[Dc],d=!!y&&("values"==y.name||null==y.name),v=ff.values;Us(u,qc,!0),Us(h,Hc,o),Us(h,Yc,!0),Us(h,Vc,u),(n?new u(1)[Gc]==o:Gc in h)||pc(h,Gc,{get:function(){return o}}),l[o]=u,Fs(Fs.G+Fs.W+Fs.F*(u!=s),l),Fs(Fs.S,o,{BYTES_PER_ELEMENT:e}),Fs(Fs.S+Fs.F*Ls((function(){s.of.call(u,1)})),o,{from:rf,of:nf}),"BYTES_PER_ELEMENT"in h||Us(h,"BYTES_PER_ELEMENT",e),Fs(Fs.P,o,uf),sc(o),Fs(Fs.P+Fs.F*zc,o,{set:cf}),Fs(Fs.P+Fs.F*!d,o,ff),h.toString!=Uc&&(h.toString=Uc),Fs(Fs.P+Fs.F*Ls((function(){new u(1).slice()})),o,{slice:sf}),Fs(Fs.P+Fs.F*(Ls((function(){return[1,2].toLocaleString()!=new u([1,2]).toLocaleString()}))||!Ls((function(){h.toLocaleString.call([1,2])}))),o,{toLocaleString:af}),ac[o]=d?y:v,d||Us(h,Dc,v)}}else us.exports=function(){};us.exports("Uint8",1,(function(t){return function(e,r,n){return t(this,e,r,n)}}));var df=y,vf=ye,gf=Ct,bf=Bt,mf=nu,wf=au,xf=Math.max,_f=Math.min,Sf=Math.floor,Ef=/\$([$&`']|\d\d?|<[^>]*>)/g,Af=/\$([$&`']|\d\d?)/g;Su("replace",2,(function(t,e,r,n){return[function(n,o){var i=t(this),a=null==n?void 0:n[e];return void 0!==a?a.call(n,i,o):r.call(String(i),n,o)},function(t,e){var i=n(r,t,this,e);if(i.done)return i.value;var a=df(t),u=String(this),s="function"==typeof e;s||(e=String(e));var c=a.global;if(c){var f=a.unicode;a.lastIndex=0}for(var l=[];;){var h=wf(a,u);if(null===h)break;if(l.push(h),!c)break;""===String(h[0])&&(a.lastIndex=mf(u,gf(a.lastIndex),f))}for(var p,y="",d=0,v=0;v<l.length;v++){h=l[v];for(var g=String(h[0]),b=xf(_f(bf(h.index),u.length),0),m=[],w=1;w<h.length;w++)m.push(void 0===(p=h[w])?p:String(p));var x=h.groups;if(s){var _=[g].concat(m,b,u);void 0!==x&&_.push(x);var S=String(e.apply(void 0,_))}else S=o(g,u,b,m,x,e);b>=d&&(y+=u.slice(d,b)+S,d=b+g.length)}return y+u.slice(d)}];function o(t,e,n,o,i,a){var u=n+t.length,s=o.length,c=Af;return void 0!==i&&(i=vf(i),c=Ef),r.call(a,c,(function(r,a){var c;switch(a.charAt(0)){case"$":return"$";case"&":return t;case"`":return e.slice(0,n);case"'":return e.slice(u);case"<":c=i[a.slice(1,-1)];break;default:var f=+a;if(0===f)return r;if(f>s){var l=Sf(f/10);return 0===l?r:l<=s?void 0===o[l-1]?a.charAt(1):o[l-1]+a.charAt(1):r}c=o[f-1]}return void 0===c?"":c}))}}));var Of="undefined"!=typeof globalThis&&globalThis||"undefined"!=typeof self&&self||void 0!==Of&&Of,Tf="URLSearchParams"in Of,Pf="Symbol"in Of&&"iterator"in Symbol,jf="FileReader"in Of&&"Blob"in Of&&function(){try{return new Blob,!0}catch(t){return!1}}(),If="FormData"in Of,Lf="ArrayBuffer"in Of;if(Lf)var Ff=["[object Int8Array]","[object Uint8Array]","[object Uint8ClampedArray]","[object Int16Array]","[object Uint16Array]","[object Int32Array]","[object Uint32Array]","[object Float32Array]","[object Float64Array]"],Rf=ArrayBuffer.isView||function(t){return t&&Ff.indexOf(Object.prototype.toString.call(t))>-1};function Nf(t){if("string"!=typeof t&&(t=String(t)),/[^a-z0-9\-#$%&'*+.^_`|~!]/i.test(t)||""===t)throw new TypeError('Invalid character in header field name: "'+t+'"');return t.toLowerCase()}function kf(t){return"string"!=typeof t&&(t=String(t)),t}function Bf(t){var e={next:function(){var e=t.shift();return{done:void 0===e,value:e}}};return Pf&&(e[Symbol.iterator]=function(){return e}),e}function Mf(t){this.map={},t instanceof Mf?t.forEach((function(t,e){this.append(e,t)}),this):Array.isArray(t)?t.forEach((function(t){this.append(t[0],t[1])}),this):t&&Object.getOwnPropertyNames(t).forEach((function(e){this.append(e,t[e])}),this)}function Uf(t){if(t.bodyUsed)return Promise.reject(new TypeError("Already read"));t.bodyUsed=!0}function Cf(t){return new Promise((function(e,r){t.onload=function(){e(t.result)},t.onerror=function(){r(t.error)}}))}function Df(t){var e=new FileReader,r=Cf(e);return e.readAsArrayBuffer(t),r}function Gf(t){if(t.slice)return t.slice(0);var e=new Uint8Array(t.byteLength);return e.set(new Uint8Array(t)),e.buffer}function qf(){return this.bodyUsed=!1,this._initBody=function(t){var e;this.bodyUsed=this.bodyUsed,this._bodyInit=t,t?"string"==typeof t?this._bodyText=t:jf&&Blob.prototype.isPrototypeOf(t)?this._bodyBlob=t:If&&FormData.prototype.isPrototypeOf(t)?this._bodyFormData=t:Tf&&URLSearchParams.prototype.isPrototypeOf(t)?this._bodyText=t.toString():Lf&&jf&&((e=t)&&DataView.prototype.isPrototypeOf(e))?(this._bodyArrayBuffer=Gf(t.buffer),this._bodyInit=new Blob([this._bodyArrayBuffer])):Lf&&(ArrayBuffer.prototype.isPrototypeOf(t)||Rf(t))?this._bodyArrayBuffer=Gf(t):this._bodyText=t=Object.prototype.toString.call(t):this._bodyText="",this.headers.get("content-type")||("string"==typeof t?this.headers.set("content-type","text/plain;charset=UTF-8"):this._bodyBlob&&this._bodyBlob.type?this.headers.set("content-type",this._bodyBlob.type):Tf&&URLSearchParams.prototype.isPrototypeOf(t)&&this.headers.set("content-type","application/x-www-form-urlencoded;charset=UTF-8"))},jf&&(this.blob=function(){var t=Uf(this);if(t)return t;if(this._bodyBlob)return Promise.resolve(this._bodyBlob);if(this._bodyArrayBuffer)return Promise.resolve(new Blob([this._bodyArrayBuffer]));if(this._bodyFormData)throw new Error("could not read FormData body as blob");return Promise.resolve(new Blob([this._bodyText]))},this.arrayBuffer=function(){if(this._bodyArrayBuffer){var t=Uf(this);return t||(ArrayBuffer.isView(this._bodyArrayBuffer)?Promise.resolve(this._bodyArrayBuffer.buffer.slice(this._bodyArrayBuffer.byteOffset,this._bodyArrayBuffer.byteOffset+this._bodyArrayBuffer.byteLength)):Promise.resolve(this._bodyArrayBuffer))}return this.blob().then(Df)}),this.text=function(){var t,e,r,n=Uf(this);if(n)return n;if(this._bodyBlob)return t=this._bodyBlob,e=new FileReader,r=Cf(e),e.readAsText(t),r;if(this._bodyArrayBuffer)return Promise.resolve(function(t){for(var e=new Uint8Array(t),r=new Array(e.length),n=0;n<e.length;n++)r[n]=String.fromCharCode(e[n]);return r.join("")}(this._bodyArrayBuffer));if(this._bodyFormData)throw new Error("could not read FormData body as text");return Promise.resolve(this._bodyText)},If&&(this.formData=function(){return this.text().then(Hf)}),this.json=function(){return this.text().then(JSON.parse)},this}Mf.prototype.append=function(t,e){t=Nf(t),e=kf(e);var r=this.map[t];this.map[t]=r?r+", "+e:e},Mf.prototype.delete=function(t){delete this.map[Nf(t)]},Mf.prototype.get=function(t){return t=Nf(t),this.has(t)?this.map[t]:null},Mf.prototype.has=function(t){return this.map.hasOwnProperty(Nf(t))},Mf.prototype.set=function(t,e){this.map[Nf(t)]=kf(e)},Mf.prototype.forEach=function(t,e){for(var r in this.map)this.map.hasOwnProperty(r)&&t.call(e,this.map[r],r,this)},Mf.prototype.keys=function(){var t=[];return this.forEach((function(e,r){t.push(r)})),Bf(t)},Mf.prototype.values=function(){var t=[];return this.forEach((function(e){t.push(e)})),Bf(t)},Mf.prototype.entries=function(){var t=[];return this.forEach((function(e,r){t.push([r,e])})),Bf(t)},Pf&&(Mf.prototype[Symbol.iterator]=Mf.prototype.entries);var Vf=["DELETE","GET","HEAD","OPTIONS","POST","PUT"];function Wf(t,e){if(!(this instanceof Wf))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');var r=(e=e||{}).body;if(t instanceof Wf){if(t.bodyUsed)throw new TypeError("Already read");this.url=t.url,this.credentials=t.credentials,e.headers||(this.headers=new Mf(t.headers)),this.method=t.method,this.mode=t.mode,this.signal=t.signal,r||null==t._bodyInit||(r=t._bodyInit,t.bodyUsed=!0)}else this.url=String(t);if(this.credentials=e.credentials||this.credentials||"same-origin",!e.headers&&this.headers||(this.headers=new Mf(e.headers)),this.method=function(t){var e=t.toUpperCase();return Vf.indexOf(e)>-1?e:t}(e.method||this.method||"GET"),this.mode=e.mode||this.mode||null,this.signal=e.signal||this.signal,this.referrer=null,("GET"===this.method||"HEAD"===this.method)&&r)throw new TypeError("Body not allowed for GET or HEAD requests");if(this._initBody(r),!("GET"!==this.method&&"HEAD"!==this.method||"no-store"!==e.cache&&"no-cache"!==e.cache)){var n=/([?&])_=[^&]*/;if(n.test(this.url))this.url=this.url.replace(n,"$1_="+(new Date).getTime());else{this.url+=(/\?/.test(this.url)?"&":"?")+"_="+(new Date).getTime()}}}function Hf(t){var e=new FormData;return t.trim().split("&").forEach((function(t){if(t){var r=t.split("="),n=r.shift().replace(/\+/g," "),o=r.join("=").replace(/\+/g," ");e.append(decodeURIComponent(n),decodeURIComponent(o))}})),e}function Yf(t,e){if(!(this instanceof Yf))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');e||(e={}),this.type="default",this.status=void 0===e.status?200:e.status,this.ok=this.status>=200&&this.status<300,this.statusText=void 0===e.statusText?"":""+e.statusText,this.headers=new Mf(e.headers),this.url=e.url||"",this._initBody(t)}Wf.prototype.clone=function(){return new Wf(this,{body:this._bodyInit})},qf.call(Wf.prototype),qf.call(Yf.prototype),Yf.prototype.clone=function(){return new Yf(this._bodyInit,{status:this.status,statusText:this.statusText,headers:new Mf(this.headers),url:this.url})},Yf.error=function(){var t=new Yf(null,{status:0,statusText:""});return t.type="error",t};var $f=[301,302,303,307,308];Yf.redirect=function(t,e){if(-1===$f.indexOf(e))throw new RangeError("Invalid status code");return new Yf(null,{status:e,headers:{location:t}})};var Jf,zf,Xf=Of.DOMException;try{new Xf}catch(t){(Xf=function(t,e){this.message=t,this.name=e;var r=Error(t);this.stack=r.stack}).prototype=Object.create(Error.prototype),Xf.prototype.constructor=Xf}function Kf(t,e){return new Promise((function(n,o){var i=new Wf(t,e);if(i.signal&&i.signal.aborted)return o(new Xf("Aborted","AbortError"));var a=new XMLHttpRequest;function u(){a.abort()}a.onload=function(){var t,e,r={status:a.status,statusText:a.statusText,headers:(t=a.getAllResponseHeaders()||"",e=new Mf,t.replace(/\r?\n[\t ]+/g," ").split("\r").map((function(t){return 0===t.indexOf("\n")?t.substr(1,t.length):t})).forEach((function(t){var r=t.split(":"),n=r.shift().trim();if(n){var o=r.join(":").trim();e.append(n,o)}})),e)};r.url="responseURL"in a?a.responseURL:r.headers.get("X-Request-URL");var o="response"in a?a.response:a.responseText;setTimeout((function(){n(new Yf(o,r))}),0)},a.onerror=function(){setTimeout((function(){o(new TypeError("Network request failed"))}),0)},a.ontimeout=function(){setTimeout((function(){o(new TypeError("Network request failed"))}),0)},a.onabort=function(){setTimeout((function(){o(new Xf("Aborted","AbortError"))}),0)},a.open(i.method,function(t){try{return""===t&&Of.location.href?Of.location.href:t}catch(e){return t}}(i.url),!0),"include"===i.credentials?a.withCredentials=!0:"omit"===i.credentials&&(a.withCredentials=!1),"responseType"in a&&(jf?a.responseType="blob":Lf&&i.headers.get("Content-Type")&&-1!==i.headers.get("Content-Type").indexOf("application/octet-stream")&&(a.responseType="arraybuffer")),!e||"object"!==r(e.headers)||e.headers instanceof Mf?i.headers.forEach((function(t,e){a.setRequestHeader(e,t)})):Object.getOwnPropertyNames(e.headers).forEach((function(t){a.setRequestHeader(t,kf(e.headers[t]))})),i.signal&&(i.signal.addEventListener("abort",u),a.onreadystatechange=function(){4===a.readyState&&i.signal.removeEventListener("abort",u)}),a.send(void 0===i._bodyInit?null:i._bodyInit)}))}function Qf(t,e){var r="undefined"!=typeof Symbol&&t[Symbol.iterator]||t["@@iterator"];if(!r){if(Array.isArray(t)||(r=function(t,e){if(!t)return;if("string"==typeof t)return Zf(t,e);var r=Object.prototype.toString.call(t).slice(8,-1);"Object"===r&&t.constructor&&(r=t.constructor.name);if("Map"===r||"Set"===r)return Array.from(t);if("Arguments"===r||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(r))return Zf(t,e)}(t))||e&&t&&"number"==typeof t.length){r&&(t=r);var n=0,o=function(){};return{s:o,n:function(){return n>=t.length?{done:!0}:{done:!1,value:t[n++]}},e:function(t){throw t},f:o}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var i,a=!0,u=!1;return{s:function(){r=r.call(t)},n:function(){var t=r.next();return a=t.done,t},e:function(t){u=!0,i=t},f:function(){try{a||null==r.return||r.return()}finally{if(u)throw i}}}}function Zf(t,e){(null==e||e>t.length)&&(e=t.length);for(var r=0,n=new Array(e);r<e;r++)n[r]=t[r];return n}Kf.polyfill=!0,Of.fetch||(Of.fetch=Kf,Of.Headers=Mf,Of.Request=Wf,Of.Response=Yf),"object"!==r(window.ui)?((Jf={}).resolveStorage={},Jf.requestCounter=0,Jf.initStarted=!1,Jf.initFinished=!1,Jf.initFailed=!1,window.ui=Jf,window.ui.copyright="© Copyright 2021, by Marco Mengelkoch"):Jf=window.ui,"object"!==r(window.backend)?(zf={},window.backend=zf):zf=window.backend;Jf.createRequestId=function(){return Jf.requestCounter>=Number.MAX_SAFE_INTEGER-1&&(Jf.requestCounter=0),Jf.requestCounter++},Jf.alert=function(t){void 0===window.nimview?alert(t):"function"==typeof window.nimview.alert&&window.nimview.alert(t)},Jf.callRequest=function(){var t=e(da.mark((function t(e,r,n){var o,i,a,u,s,c;return da.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(void 0===r&&(r=""),void 0===n&&(n=[]),void 0!==window.nimview){t.next=12;break}return o=JSON.stringify({request:e,data:n}),i={method:"POST",mode:"cors",cache:"no-cache",headers:{"Content-Type":"application/json"},body:o},a=e,Jf.globalToken&&Jf.globalToken.length>0&&(i.headers["global-token"]=Jf.globalToken),n.length!=Math.min(r.length,r.split(",").length)&&console&&console.log&&console.log(0&e&r&"' data: '"&JSON.stringify(n)&"'"),t.abrupt("return",fetch("/"+a,i).then((function(t){if(t){var e=t.headers.get("global-token");e&&e.length>0&&(Jf.globalToken=e,Jf.lastToken=Date.now())}return t.text()})));case 12:if("function"!=typeof window.nimview.call){t.next=17;break}return u=Jf.createRequestId(),s=JSON.stringify({request:e,data:n,requestId:u}),c=new Promise((function(t,e){Jf.resolveStorage[u]=[t,e];var r=window.nimview.call(s);void 0!==r&&(t(r),delete Jf.resolveStorage[u])})),t.abrupt("return",c);case 17:case"end":return t.stop()}}),t)})));return function(e,r,n){return t.apply(this,arguments)}}(),Jf.addRequest=function(t){"string"==typeof t&&(t=[t,""]);var r=t[0]+"",n=t[1]+"";zf[r]=e(da.mark((function t(){var e,o,i,a=arguments;return da.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:for(e=a.length,o=new Array(e),i=0;i<e;i++)o[i]=a[i];return t.abrupt("return",Jf.callRequest(r,n,o));case 2:case"end":return t.stop()}}),t)})))},Jf.applyResponse=function(t,e){void 0!==Jf.resolveStorage[t]?(Jf.resolveStorage[t][0](e),delete Jf.resolveStorage[t]):Jf.alert("request id '"+t+"' not found")},Jf.rejectResponse=function(t){void 0!==Jf.resolveStorage[t]&&(Jf.resolveStorage[t][1](),delete Jf.resolveStorage[t])},Jf.initRequests=e(da.mark((function t(){return da.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(0!=Jf.initStarted){t.next=6;break}return Jf.initStarted=!0,t.next=4,Jf.callRequest("getGlobalToken","",[]).then((function(t){JSON.parse(t).useGlobalToken&&window.setInterval(Jf.getGlobalToken,6e4)})).catch((function(){Jf.alert("getGlobalToken failed"),Jf.initFailed=!0}));case 4:return t.next=6,Jf.callRequest("getRequests","",[]).then((function(t){var e,r=Qf(JSON.parse(t));try{for(r.s();!(e=r.n()).done;){var n=e.value;Jf.addRequest(n)}}catch(t){r.e(t)}finally{r.f()}Jf.initFinished=!0})).catch((function(){Jf.alert("getRequests failed"),Jf.initFailed=!0}));case 6:case"end":return t.stop()}}),t)}))),zf.waitInit=function(){return new Promise((function(t,e){setTimeout((function r(){return Jf.initFinished?t():Jf.initFailed?e():void setTimeout(r,5)}),1)}))},window.setTimeout(Jf.initRequests,3);var tl=zf;module.exports=tl;
+//# sourceMappingURL=nimview.cjs.js.map
