+++ src/js/nimview.iife.js
@@ -1,0 +1,2 @@
+var nimview_js=function(){"use strict";function t(t,e,r,n,o,i,a){try{var u=t[i](a),s=u.value}catch(t){return void r(t)}u.done?e(s):Promise.resolve(s).then(n,o)}function e(e){return function(){var r=this,n=arguments;return new Promise((function(o,i){var a=e.apply(r,n);function u(e){t(a,o,i,u,s,"next",e)}function s(e){t(a,o,i,u,s,"throw",e)}u(void 0)}))}}function r(t){return(r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(t){return typeof t}:function(t){return t&&"function"==typeof Symbol&&t.constructor===Symbol&&t!==Symbol.prototype?"symbol":typeof t})(t)}var n={exports:{}},o=n.exports="undefined"!=typeof window&&window.Math==Math?window:"undefined"!=typeof self&&self.Math==Math?self:Function("return this")();"number"==typeof __g&&(__g=o);var i={}.hasOwnProperty,a=function(t,e){return i.call(t,e)},u=function(t){try{return!!t()}catch(t){return!0}},s=!u((function(){return 7!=Object.defineProperty({},"a",{get:function(){return 7}}).a})),c={exports:{}},f=c.exports={version:"2.6.12"};"number"==typeof __e&&(__e=f);var l={},h=function(t){return"object"==typeof t?null!==t:"function"==typeof t},p=h,y=function(t){if(!p(t))throw TypeError(t+" is not an object!");return t},d=h,v=n.exports.document,g=d(v)&&d(v.createElement),b=function(t){return g?v.createElement(t):{}},m=!s&&!u((function(){return 7!=Object.defineProperty(b("div"),"a",{get:function(){return 7}}).a})),w=h,x=function(t,e){if(!w(t))return t;var r,n;if(e&&"function"==typeof(r=t.toString)&&!w(n=r.call(t)))return n;if("function"==typeof(r=t.valueOf)&&!w(n=r.call(t)))return n;if(!e&&"function"==typeof(r=t.toString)&&!w(n=r.call(t)))return n;throw TypeError("Can't convert object to primitive value")},_=y,S=m,E=x,O=Object.defineProperty;l.f=s?Object.defineProperty:function(t,e,r){if(_(t),e=E(e,!0),_(r),S)try{return O(t,e,r)}catch(t){}if("get"in r||"set"in r)throw TypeError("Accessors not supported!");return"value"in r&&(t[e]=r.value),t};var A=function(t,e){return{enumerable:!(1&t),configurable:!(2&t),writable:!(4&t),value:e}},T=l,j=A,P=s?function(t,e,r){return T.f(t,e,j(1,r))}:function(t,e,r){return t[e]=r,t},I={exports:{}},L=0,F=Math.random(),R=function(t){return"Symbol(".concat(void 0===t?"":t,")_",(++L+F).toString(36))},N={exports:{}},k=c.exports,B=n.exports,M="__core-js_shared__",U=B[M]||(B[M]={});(N.exports=function(t,e){return U[t]||(U[t]=void 0!==e?e:{})})("versions",[]).push({version:k.version,mode:"global",copyright:"© 2020 Denis Pushkarev (zloirock.ru)"});var C=N.exports("native-function-to-string",Function.toString),D=n.exports,G=P,q=a,V=R("src"),W=C,H="toString",$=(""+W).split(H);c.exports.inspectSource=function(t){return W.call(t)},(I.exports=function(t,e,r,n){var o="function"==typeof r;o&&(q(r,"name")||G(r,"name",e)),t[e]!==r&&(o&&(q(r,V)||G(r,V,t[e]?""+t[e]:$.join(String(e)))),t===D?t[e]=r:n?t[e]?t[e]=r:G(t,e,r):(delete t[e],G(t,e,r)))})(Function.prototype,H,(function(){return"function"==typeof this&&this[V]||W.call(this)}));var Y=function(t){if("function"!=typeof t)throw TypeError(t+" is not a function!");return t},J=Y,z=function(t,e,r){if(J(t),void 0===e)return t;switch(r){case 1:return function(r){return t.call(e,r)};case 2:return function(r,n){return t.call(e,r,n)};case 3:return function(r,n,o){return t.call(e,r,n,o)}}return function(){return t.apply(e,arguments)}},X=n.exports,K=c.exports,Q=P,Z=I.exports,tt=z,et=function(t,e,r){var n,o,i,a,u=t&et.F,s=t&et.G,c=t&et.S,f=t&et.P,l=t&et.B,h=s?X:c?X[e]||(X[e]={}):(X[e]||{}).prototype,p=s?K:K[e]||(K[e]={}),y=p.prototype||(p.prototype={});for(n in s&&(r=e),r)i=((o=!u&&h&&void 0!==h[n])?h:r)[n],a=l&&o?tt(i,X):f&&"function"==typeof i?tt(Function.call,i):i,h&&Z(h,n,i,t&et.U),p[n]!=i&&Q(p,n,a),f&&y[n]!=i&&(y[n]=i)};X.core=K,et.F=1,et.G=2,et.S=4,et.P=8,et.B=16,et.W=32,et.U=64,et.R=128;var rt=et,nt={exports:{}},ot=R("meta"),it=h,at=a,ut=l.f,st=0,ct=Object.isExtensible||function(){return!0},ft=!u((function(){return ct(Object.preventExtensions({}))})),lt=function(t){ut(t,ot,{value:{i:"O"+ ++st,w:{}}})},ht=nt.exports={KEY:ot,NEED:!1,fastKey:function(t,e){if(!it(t))return"symbol"==typeof t?t:("string"==typeof t?"S":"P")+t;if(!at(t,ot)){if(!ct(t))return"F";if(!e)return"E";lt(t)}return t[ot].i},getWeak:function(t,e){if(!at(t,ot)){if(!ct(t))return!0;if(!e)return!1;lt(t)}return t[ot].w},onFreeze:function(t){return ft&&ht.NEED&&ct(t)&&!at(t,ot)&&lt(t),t}},pt={exports:{}},yt=N.exports("wks"),dt=R,vt=n.exports.Symbol,gt="function"==typeof vt;(pt.exports=function(t){return yt[t]||(yt[t]=gt&&vt[t]||(gt?vt:dt)("Symbol."+t))}).store=yt;var bt=l.f,mt=a,wt=pt.exports("toStringTag"),xt=function(t,e,r){t&&!mt(t=r?t:t.prototype,wt)&&bt(t,wt,{configurable:!0,value:e})},_t={};_t.f=pt.exports;var St=n.exports,Et=c.exports,Ot=_t,At=l.f,Tt=function(t){var e=Et.Symbol||(Et.Symbol=St.Symbol||{});"_"==t.charAt(0)||t in e||At(e,t,{value:Ot.f(t)})},jt={}.toString,Pt=function(t){return jt.call(t).slice(8,-1)},It=Pt,Lt=Object("z").propertyIsEnumerable(0)?Object:function(t){return"String"==It(t)?t.split(""):Object(t)},Ft=function(t){if(null==t)throw TypeError("Can't call method on  "+t);return t},Rt=Lt,Nt=Ft,kt=function(t){return Rt(Nt(t))},Bt=Math.ceil,Mt=Math.floor,Ut=function(t){return isNaN(t=+t)?0:(t>0?Mt:Bt)(t)},Ct=Ut,Dt=Math.min,Gt=function(t){return t>0?Dt(Ct(t),9007199254740991):0},qt=Ut,Vt=Math.max,Wt=Math.min,Ht=function(t,e){return(t=qt(t))<0?Vt(t+e,0):Wt(t,e)},$t=kt,Yt=Gt,Jt=Ht,zt=function(t){return function(e,r,n){var o,i=$t(e),a=Yt(i.length),u=Jt(n,a);if(t&&r!=r){for(;a>u;)if((o=i[u++])!=o)return!0}else for(;a>u;u++)if((t||u in i)&&i[u]===r)return t||u||0;return!t&&-1}},Xt=N.exports("keys"),Kt=R,Qt=function(t){return Xt[t]||(Xt[t]=Kt(t))},Zt=a,te=kt,ee=zt(!1),re=Qt("IE_PROTO"),ne=function(t,e){var r,n=te(t),o=0,i=[];for(r in n)r!=re&&Zt(n,r)&&i.push(r);for(;e.length>o;)Zt(n,r=e[o++])&&(~ee(i,r)||i.push(r));return i},oe="constructor,hasOwnProperty,isPrototypeOf,propertyIsEnumerable,toLocaleString,toString,valueOf".split(","),ie=ne,ae=oe,ue=Object.keys||function(t){return ie(t,ae)},se={};se.f=Object.getOwnPropertySymbols;var ce={};ce.f={}.propertyIsEnumerable;var fe=ue,le=se,he=ce,pe=Pt,ye=Array.isArray||function(t){return"Array"==pe(t)},de=Ft,ve=function(t){return Object(de(t))},ge=l,be=y,me=ue,we=s?Object.defineProperties:function(t,e){be(t);for(var r,n=me(e),o=n.length,i=0;o>i;)ge.f(t,r=n[i++],e[r]);return t},xe=n.exports.document,_e=xe&&xe.documentElement,Se=y,Ee=we,Oe=oe,Ae=Qt("IE_PROTO"),Te=function(){},je=function(){var t,e=b("iframe"),r=Oe.length;for(e.style.display="none",_e.appendChild(e),e.src="javascript:",(t=e.contentWindow.document).open(),t.write("<script>document.F=Object<\/script>"),t.close(),je=t.F;r--;)delete je.prototype[Oe[r]];return je()},Pe=Object.create||function(t,e){var r;return null!==t?(Te.prototype=Se(t),r=new Te,Te.prototype=null,r[Ae]=t):r=je(),void 0===e?r:Ee(r,e)},Ie={},Le={},Fe=ne,Re=oe.concat("length","prototype");Le.f=Object.getOwnPropertyNames||function(t){return Fe(t,Re)};var Ne=kt,ke=Le.f,Be={}.toString,Me="object"==typeof window&&window&&Object.getOwnPropertyNames?Object.getOwnPropertyNames(window):[];Ie.f=function(t){return Me&&"[object Window]"==Be.call(t)?function(t){try{return ke(t)}catch(t){return Me.slice()}}(t):ke(Ne(t))};var Ue={},Ce=ce,De=A,Ge=kt,qe=x,Ve=a,We=m,He=Object.getOwnPropertyDescriptor;Ue.f=s?He:function(t,e){if(t=Ge(t),e=qe(e,!0),We)try{return He(t,e)}catch(t){}if(Ve(t,e))return De(!Ce.f.call(t,e),t[e])};var $e=n.exports,Ye=a,Je=s,ze=rt,Xe=I.exports,Ke=nt.exports.KEY,Qe=u,Ze=N.exports,tr=xt,er=R,rr=pt.exports,nr=_t,or=Tt,ir=function(t){var e=fe(t),r=le.f;if(r)for(var n,o=r(t),i=he.f,a=0;o.length>a;)i.call(t,n=o[a++])&&e.push(n);return e},ar=ye,ur=y,sr=h,cr=ve,fr=kt,lr=x,hr=A,pr=Pe,yr=Ie,dr=Ue,vr=se,gr=l,br=ue,mr=dr.f,wr=gr.f,xr=yr.f,_r=$e.Symbol,Sr=$e.JSON,Er=Sr&&Sr.stringify,Or=rr("_hidden"),Ar=rr("toPrimitive"),Tr={}.propertyIsEnumerable,jr=Ze("symbol-registry"),Pr=Ze("symbols"),Ir=Ze("op-symbols"),Lr=Object.prototype,Fr="function"==typeof _r&&!!vr.f,Rr=$e.QObject,Nr=!Rr||!Rr.prototype||!Rr.prototype.findChild,kr=Je&&Qe((function(){return 7!=pr(wr({},"a",{get:function(){return wr(this,"a",{value:7}).a}})).a}))?function(t,e,r){var n=mr(Lr,e);n&&delete Lr[e],wr(t,e,r),n&&t!==Lr&&wr(Lr,e,n)}:wr,Br=function(t){var e=Pr[t]=pr(_r.prototype);return e._k=t,e},Mr=Fr&&"symbol"==typeof _r.iterator?function(t){return"symbol"==typeof t}:function(t){return t instanceof _r},Ur=function(t,e,r){return t===Lr&&Ur(Ir,e,r),ur(t),e=lr(e,!0),ur(r),Ye(Pr,e)?(r.enumerable?(Ye(t,Or)&&t[Or][e]&&(t[Or][e]=!1),r=pr(r,{enumerable:hr(0,!1)})):(Ye(t,Or)||wr(t,Or,hr(1,{})),t[Or][e]=!0),kr(t,e,r)):wr(t,e,r)},Cr=function(t,e){ur(t);for(var r,n=ir(e=fr(e)),o=0,i=n.length;i>o;)Ur(t,r=n[o++],e[r]);return t},Dr=function(t){var e=Tr.call(this,t=lr(t,!0));return!(this===Lr&&Ye(Pr,t)&&!Ye(Ir,t))&&(!(e||!Ye(this,t)||!Ye(Pr,t)||Ye(this,Or)&&this[Or][t])||e)},Gr=function(t,e){if(t=fr(t),e=lr(e,!0),t!==Lr||!Ye(Pr,e)||Ye(Ir,e)){var r=mr(t,e);return!r||!Ye(Pr,e)||Ye(t,Or)&&t[Or][e]||(r.enumerable=!0),r}},qr=function(t){for(var e,r=xr(fr(t)),n=[],o=0;r.length>o;)Ye(Pr,e=r[o++])||e==Or||e==Ke||n.push(e);return n},Vr=function(t){for(var e,r=t===Lr,n=xr(r?Ir:fr(t)),o=[],i=0;n.length>i;)!Ye(Pr,e=n[i++])||r&&!Ye(Lr,e)||o.push(Pr[e]);return o};Fr||(Xe((_r=function(){if(this instanceof _r)throw TypeError("Symbol is not a constructor!");var t=er(arguments.length>0?arguments[0]:void 0),e=function(r){this===Lr&&e.call(Ir,r),Ye(this,Or)&&Ye(this[Or],t)&&(this[Or][t]=!1),kr(this,t,hr(1,r))};return Je&&Nr&&kr(Lr,t,{configurable:!0,set:e}),Br(t)}).prototype,"toString",(function(){return this._k})),dr.f=Gr,gr.f=Ur,Le.f=yr.f=qr,ce.f=Dr,vr.f=Vr,Je&&Xe(Lr,"propertyIsEnumerable",Dr,!0),nr.f=function(t){return Br(rr(t))}),ze(ze.G+ze.W+ze.F*!Fr,{Symbol:_r});for(var Wr="hasInstance,isConcatSpreadable,iterator,match,replace,search,species,split,toPrimitive,toStringTag,unscopables".split(","),Hr=0;Wr.length>Hr;)rr(Wr[Hr++]);for(var $r=br(rr.store),Yr=0;$r.length>Yr;)or($r[Yr++]);ze(ze.S+ze.F*!Fr,"Symbol",{for:function(t){return Ye(jr,t+="")?jr[t]:jr[t]=_r(t)},keyFor:function(t){if(!Mr(t))throw TypeError(t+" is not a symbol!");for(var e in jr)if(jr[e]===t)return e},useSetter:function(){Nr=!0},useSimple:function(){Nr=!1}}),ze(ze.S+ze.F*!Fr,"Object",{create:function(t,e){return void 0===e?pr(t):Cr(pr(t),e)},defineProperty:Ur,defineProperties:Cr,getOwnPropertyDescriptor:Gr,getOwnPropertyNames:qr,getOwnPropertySymbols:Vr});var Jr=Qe((function(){vr.f(1)}));ze(ze.S+ze.F*Jr,"Object",{getOwnPropertySymbols:function(t){return vr.f(cr(t))}}),Sr&&ze(ze.S+ze.F*(!Fr||Qe((function(){var t=_r();return"[null]"!=Er([t])||"{}"!=Er({a:t})||"{}"!=Er(Object(t))}))),"JSON",{stringify:function(t){for(var e,r,n=[t],o=1;arguments.length>o;)n.push(arguments[o++]);if(r=e=n[1],(sr(e)||void 0!==t)&&!Mr(t))return ar(e)||(e=function(t,e){if("function"==typeof r&&(e=r.call(this,t,e)),!Mr(e))return e}),n[1]=e,Er.apply(Sr,n)}}),_r.prototype[Ar]||P(_r.prototype,Ar,_r.prototype.valueOf),tr(_r,"Symbol"),tr(Math,"Math",!0),tr($e.JSON,"JSON",!0);var zr=Ut,Xr=Ft,Kr=function(t){return function(e,r){var n,o,i=String(Xr(e)),a=zr(r),u=i.length;return a<0||a>=u?t?"":void 0:(n=i.charCodeAt(a))<55296||n>56319||a+1===u||(o=i.charCodeAt(a+1))<56320||o>57343?t?i.charAt(a):n:t?i.slice(a,a+2):o-56320+(n-55296<<10)+65536}},Qr={},Zr=Pe,tn=A,en=xt,rn={};P(rn,pt.exports("iterator"),(function(){return this}));var nn=a,on=ve,an=Qt("IE_PROTO"),un=Object.prototype,sn=Object.getPrototypeOf||function(t){return t=on(t),nn(t,an)?t[an]:"function"==typeof t.constructor&&t instanceof t.constructor?t.constructor.prototype:t instanceof Object?un:null},cn=rt,fn=I.exports,ln=P,hn=Qr,pn=function(t,e,r){t.prototype=Zr(rn,{next:tn(1,r)}),en(t,e+" Iterator")},yn=xt,dn=sn,vn=pt.exports("iterator"),gn=!([].keys&&"next"in[].keys()),bn="keys",mn="values",wn=function(){return this},xn=function(t,e,r,n,o,i,a){pn(r,e,n);var u,s,c,f=function(t){if(!gn&&t in y)return y[t];switch(t){case bn:case mn:return function(){return new r(this,t)}}return function(){return new r(this,t)}},l=e+" Iterator",h=o==mn,p=!1,y=t.prototype,d=y[vn]||y["@@iterator"]||o&&y[o],v=d||f(o),g=o?h?f("entries"):v:void 0,b="Array"==e&&y.entries||d;if(b&&(c=dn(b.call(new t)))!==Object.prototype&&c.next&&(yn(c,l,!0),"function"!=typeof c[vn]&&ln(c,vn,wn)),h&&d&&d.name!==mn&&(p=!0,v=function(){return d.call(this)}),(gn||p||!y[vn])&&ln(y,vn,v),hn[e]=v,hn[l]=wn,o)if(u={values:h?v:f(mn),keys:i?v:f(bn),entries:g},a)for(s in u)s in y||fn(y,s,u[s]);else cn(cn.P+cn.F*(gn||p),e,u);return u},_n=Kr(!0);xn(String,"String",(function(t){this._t=String(t),this._i=0}),(function(){var t,e=this._t,r=this._i;return r>=e.length?{value:void 0,done:!0}:(t=_n(e,r),this._i+=t.length,{value:t,done:!1})}));var Sn=Pt,En=pt.exports("toStringTag"),On="Arguments"==Sn(function(){return arguments}()),An=function(t){var e,r,n;return void 0===t?"Undefined":null===t?"Null":"string"==typeof(r=function(t,e){try{return t[e]}catch(t){}}(e=Object(t),En))?r:On?Sn(e):"Object"==(n=Sn(e))&&"function"==typeof e.callee?"Arguments":n},Tn=An,jn={};jn[pt.exports("toStringTag")]="z",jn+""!="[object z]"&&I.exports(Object.prototype,"toString",(function(){return"[object "+Tn(this)+"]"}),!0);var Pn=pt.exports("unscopables"),In=Array.prototype;null==In[Pn]&&P(In,Pn,{});var Ln=function(t){In[Pn][t]=!0},Fn=function(t,e){return{value:e,done:!!t}},Rn=Qr,Nn=kt,kn=xn(Array,"Array",(function(t,e){this._t=Nn(t),this._i=0,this._k=e}),(function(){var t=this._t,e=this._k,r=this._i++;return!t||r>=t.length?(this._t=void 0,Fn(1)):Fn(0,"keys"==e?r:"values"==e?t[r]:[r,t[r]])}),"values");Rn.Arguments=Rn.Array,Ln("keys"),Ln("values"),Ln("entries");for(var Bn=kn,Mn=ue,Un=I.exports,Cn=n.exports,Dn=P,Gn=Qr,qn=pt.exports,Vn=qn("iterator"),Wn=qn("toStringTag"),Hn=Gn.Array,$n={CSSRuleList:!0,CSSStyleDeclaration:!1,CSSValueList:!1,ClientRectList:!1,DOMRectList:!1,DOMStringList:!1,DOMTokenList:!0,DataTransferItemList:!1,FileList:!1,HTMLAllCollection:!1,HTMLCollection:!1,HTMLFormElement:!1,HTMLSelectElement:!1,MediaList:!0,MimeTypeArray:!1,NamedNodeMap:!1,NodeList:!0,PaintRequestList:!1,Plugin:!1,PluginArray:!1,SVGLengthList:!1,SVGNumberList:!1,SVGPathSegList:!1,SVGPointList:!1,SVGStringList:!1,SVGTransformList:!1,SourceBufferList:!1,StyleSheetList:!0,TextTrackCueList:!1,TextTrackList:!1,TouchList:!1},Yn=Mn($n),Jn=0;Jn<Yn.length;Jn++){var zn,Xn=Yn[Jn],Kn=$n[Xn],Qn=Cn[Xn],Zn=Qn&&Qn.prototype;if(Zn&&(Zn[Vn]||Dn(Zn,Vn,Hn),Zn[Wn]||Dn(Zn,Wn,Xn),Gn[Xn]=Hn,Kn))for(zn in Bn)Zn[zn]||Un(Zn,zn,Bn[zn],!0)}Tt("asyncIterator");var to=rt,eo=c.exports,ro=u,no=function(t,e){var r=(eo.Object||{})[t]||Object[t],n={};n[t]=e(r),to(to.S+to.F*ro((function(){r(1)})),"Object",n)},oo=ve,io=sn;no("getPrototypeOf",(function(){return function(t){return io(oo(t))}}));var ao=l.f,uo=Function.prototype,so=/^\s*function ([^ (]*)/,co="name";co in uo||s&&ao(uo,co,{configurable:!0,get:function(){try{return(""+this).match(so)[1]}catch(t){return""}}});var fo=function(t,e,r,n){if(!(t instanceof e)||void 0!==n&&n in t)throw TypeError(r+": incorrect invocation!");return t},lo={exports:{}},ho=y,po=function(t,e,r,n){try{return n?e(ho(r)[0],r[1]):e(r)}catch(e){var o=t.return;throw void 0!==o&&ho(o.call(t)),e}},yo=Qr,vo=pt.exports("iterator"),go=Array.prototype,bo=function(t){return void 0!==t&&(yo.Array===t||go[vo]===t)},mo=An,wo=pt.exports("iterator"),xo=Qr,_o=c.exports.getIteratorMethod=function(t){if(null!=t)return t[wo]||t["@@iterator"]||xo[mo(t)]},So=z,Eo=po,Oo=bo,Ao=y,To=Gt,jo=_o,Po={},Io={},Lo=lo.exports=function(t,e,r,n,o){var i,a,u,s,c=o?function(){return t}:jo(t),f=So(r,n,e?2:1),l=0;if("function"!=typeof c)throw TypeError(t+" is not iterable!");if(Oo(c)){for(i=To(t.length);i>l;l++)if((s=e?f(Ao(a=t[l])[0],a[1]):f(t[l]))===Po||s===Io)return s}else for(u=c.call(t);!(a=u.next()).done;)if((s=Eo(u,f,a.value,e))===Po||s===Io)return s};Lo.BREAK=Po,Lo.RETURN=Io;var Fo,Ro,No,ko=y,Bo=Y,Mo=pt.exports("species"),Uo=function(t,e){var r,n=ko(t).constructor;return void 0===n||null==(r=ko(n)[Mo])?e:Bo(r)},Co=z,Do=function(t,e,r){var n=void 0===r;switch(e.length){case 0:return n?t():t.call(r);case 1:return n?t(e[0]):t.call(r,e[0]);case 2:return n?t(e[0],e[1]):t.call(r,e[0],e[1]);case 3:return n?t(e[0],e[1],e[2]):t.call(r,e[0],e[1],e[2]);case 4:return n?t(e[0],e[1],e[2],e[3]):t.call(r,e[0],e[1],e[2],e[3])}return t.apply(r,e)},Go=_e,qo=b,Vo=n.exports,Wo=Vo.process,Ho=Vo.setImmediate,$o=Vo.clearImmediate,Yo=Vo.MessageChannel,Jo=Vo.Dispatch,zo=0,Xo={},Ko="onreadystatechange",Qo=function(){var t=+this;if(Xo.hasOwnProperty(t)){var e=Xo[t];delete Xo[t],e()}},Zo=function(t){Qo.call(t.data)};Ho&&$o||(Ho=function(t){for(var e=[],r=1;arguments.length>r;)e.push(arguments[r++]);return Xo[++zo]=function(){Do("function"==typeof t?t:Function(t),e)},Fo(zo),zo},$o=function(t){delete Xo[t]},"process"==Pt(Wo)?Fo=function(t){Wo.nextTick(Co(Qo,t,1))}:Jo&&Jo.now?Fo=function(t){Jo.now(Co(Qo,t,1))}:Yo?(No=(Ro=new Yo).port2,Ro.port1.onmessage=Zo,Fo=Co(No.postMessage,No,1)):Vo.addEventListener&&"function"==typeof postMessage&&!Vo.importScripts?(Fo=function(t){Vo.postMessage(t+"","*")},Vo.addEventListener("message",Zo,!1)):Fo=Ko in qo("script")?function(t){Go.appendChild(qo("script")).onreadystatechange=function(){Go.removeChild(this),Qo.call(t)}}:function(t){setTimeout(Co(Qo,t,1),0)});var ti={set:Ho,clear:$o},ei=n.exports,ri=ti.set,ni=ei.MutationObserver||ei.WebKitMutationObserver,oi=ei.process,ii=ei.Promise,ai="process"==Pt(oi),ui={},si=Y;function ci(t){var e,r;this.promise=new t((function(t,n){if(void 0!==e||void 0!==r)throw TypeError("Bad Promise constructor");e=t,r=n})),this.resolve=si(e),this.reject=si(r)}ui.f=function(t){return new ci(t)};var fi=n.exports.navigator,li=fi&&fi.userAgent||"",hi=y,pi=h,yi=ui,di=I.exports,vi=function(t,e,r){for(var n in e)di(t,n,e[n],r);return t},gi=n.exports,bi=l,mi=s,wi=pt.exports("species"),xi=function(t){var e=gi[t];mi&&e&&!e[wi]&&bi.f(e,wi,{configurable:!0,get:function(){return this}})},_i=pt.exports("iterator"),Si=!1;try{var Ei=[7][_i]();Ei.return=function(){Si=!0},Array.from(Ei,(function(){throw 2}))}catch(t){}var Oi,Ai,Ti,ji,Pi=function(t,e){if(!e&&!Si)return!1;var r=!1;try{var n=[7],o=n[_i]();o.next=function(){return{done:r=!0}},n[_i]=function(){return o},t(n)}catch(t){}return r},Ii=n.exports,Li=z,Fi=An,Ri=rt,Ni=h,ki=Y,Bi=fo,Mi=lo.exports,Ui=Uo,Ci=ti.set,Di=function(){var t,e,r,n=function(){var n,o;for(ai&&(n=oi.domain)&&n.exit();t;){o=t.fn,t=t.next;try{o()}catch(n){throw t?r():e=void 0,n}}e=void 0,n&&n.enter()};if(ai)r=function(){oi.nextTick(n)};else if(!ni||ei.navigator&&ei.navigator.standalone)if(ii&&ii.resolve){var o=ii.resolve(void 0);r=function(){o.then(n)}}else r=function(){ri.call(ei,n)};else{var i=!0,a=document.createTextNode("");new ni(n).observe(a,{characterData:!0}),r=function(){a.data=i=!i}}return function(n){var o={fn:n,next:void 0};e&&(e.next=o),t||(t=o,r()),e=o}}(),Gi=ui,qi=function(t){try{return{e:!1,v:t()}}catch(t){return{e:!0,v:t}}},Vi=li,Wi=function(t,e){if(hi(t),pi(e)&&e.constructor===t)return e;var r=yi.f(t);return(0,r.resolve)(e),r.promise},Hi="Promise",$i=Ii.TypeError,Yi=Ii.process,Ji=Yi&&Yi.versions,zi=Ji&&Ji.v8||"",Xi=Ii.Promise,Ki="process"==Fi(Yi),Qi=function(){},Zi=Ai=Gi.f,ta=!!function(){try{var t=Xi.resolve(1),e=(t.constructor={})[pt.exports("species")]=function(t){t(Qi,Qi)};return(Ki||"function"==typeof PromiseRejectionEvent)&&t.then(Qi)instanceof e&&0!==zi.indexOf("6.6")&&-1===Vi.indexOf("Chrome/66")}catch(t){}}(),ea=function(t){var e;return!(!Ni(t)||"function"!=typeof(e=t.then))&&e},ra=function(t,e){if(!t._n){t._n=!0;var r=t._c;Di((function(){for(var n=t._v,o=1==t._s,i=0,a=function(e){var r,i,a,u=o?e.ok:e.fail,s=e.resolve,c=e.reject,f=e.domain;try{u?(o||(2==t._h&&ia(t),t._h=1),!0===u?r=n:(f&&f.enter(),r=u(n),f&&(f.exit(),a=!0)),r===e.promise?c($i("Promise-chain cycle")):(i=ea(r))?i.call(r,s,c):s(r)):c(n)}catch(t){f&&!a&&f.exit(),c(t)}};r.length>i;)a(r[i++]);t._c=[],t._n=!1,e&&!t._h&&na(t)}))}},na=function(t){Ci.call(Ii,(function(){var e,r,n,o=t._v,i=oa(t);if(i&&(e=qi((function(){Ki?Yi.emit("unhandledRejection",o,t):(r=Ii.onunhandledrejection)?r({promise:t,reason:o}):(n=Ii.console)&&n.error&&n.error("Unhandled promise rejection",o)})),t._h=Ki||oa(t)?2:1),t._a=void 0,i&&e.e)throw e.v}))},oa=function(t){return 1!==t._h&&0===(t._a||t._c).length},ia=function(t){Ci.call(Ii,(function(){var e;Ki?Yi.emit("rejectionHandled",t):(e=Ii.onrejectionhandled)&&e({promise:t,reason:t._v})}))},aa=function(t){var e=this;e._d||(e._d=!0,(e=e._w||e)._v=t,e._s=2,e._a||(e._a=e._c.slice()),ra(e,!0))},ua=function(t){var e,r=this;if(!r._d){r._d=!0,r=r._w||r;try{if(r===t)throw $i("Promise can't be resolved itself");(e=ea(t))?Di((function(){var n={_w:r,_d:!1};try{e.call(t,Li(ua,n,1),Li(aa,n,1))}catch(t){aa.call(n,t)}})):(r._v=t,r._s=1,ra(r,!1))}catch(t){aa.call({_w:r,_d:!1},t)}}};ta||(Xi=function(t){Bi(this,Xi,Hi,"_h"),ki(t),Oi.call(this);try{t(Li(ua,this,1),Li(aa,this,1))}catch(t){aa.call(this,t)}},(Oi=function(t){this._c=[],this._a=void 0,this._s=0,this._d=!1,this._v=void 0,this._h=0,this._n=!1}).prototype=vi(Xi.prototype,{then:function(t,e){var r=Zi(Ui(this,Xi));return r.ok="function"!=typeof t||t,r.fail="function"==typeof e&&e,r.domain=Ki?Yi.domain:void 0,this._c.push(r),this._a&&this._a.push(r),this._s&&ra(this,!1),r.promise},catch:function(t){return this.then(void 0,t)}}),Ti=function(){var t=new Oi;this.promise=t,this.resolve=Li(ua,t,1),this.reject=Li(aa,t,1)},Gi.f=Zi=function(t){return t===Xi||t===ji?new Ti(t):Ai(t)}),Ri(Ri.G+Ri.W+Ri.F*!ta,{Promise:Xi}),xt(Xi,Hi),xi(Hi),ji=c.exports.Promise,Ri(Ri.S+Ri.F*!ta,Hi,{reject:function(t){var e=Zi(this);return(0,e.reject)(t),e.promise}}),Ri(Ri.S+Ri.F*!ta,Hi,{resolve:function(t){return Wi(this,t)}}),Ri(Ri.S+Ri.F*!(ta&&Pi((function(t){Xi.all(t).catch(Qi)}))),Hi,{all:function(t){var e=this,r=Zi(e),n=r.resolve,o=r.reject,i=qi((function(){var r=[],i=0,a=1;Mi(t,!1,(function(t){var u=i++,s=!1;r.push(void 0),a++,e.resolve(t).then((function(t){s||(s=!0,r[u]=t,--a||n(r))}),o)})),--a||n(r)}));return i.e&&o(i.v),r.promise},race:function(t){var e=this,r=Zi(e),n=r.reject,o=qi((function(){Mi(t,!1,(function(t){e.resolve(t).then(r.resolve,n)}))}));return o.e&&n(o.v),r.promise}});var sa=y,ca=function(){var t=sa(this),e="";return t.global&&(e+="g"),t.ignoreCase&&(e+="i"),t.multiline&&(e+="m"),t.unicode&&(e+="u"),t.sticky&&(e+="y"),e};s&&"g"!=/./g.flags&&l.f(RegExp.prototype,"flags",{configurable:!0,get:ca});var fa=y,la=ca,ha=s,pa="toString",ya=/./.toString,da=function(t){I.exports(RegExp.prototype,pa,t,!0)};u((function(){return"/a/b"!=ya.call({source:"a",flags:"b"})}))?da((function(){var t=fa(this);return"/".concat(t.source,"/","flags"in t?t.flags:!ha&&t instanceof RegExp?la.call(t):void 0)})):ya.name!=pa&&da((function(){return ya.call(this)}));var va=rt,ga=_e,ba=Pt,ma=Ht,wa=Gt,xa=[].slice;va(va.P+va.F*u((function(){ga&&xa.call(ga)})),"Array",{slice:function(t,e){var r=wa(this.length),n=ba(this);if(e=void 0===e?r:e,"Array"==n)return xa.call(this,t,e);for(var o=ma(t,r),i=ma(e,r),a=wa(i-o),u=new Array(a),s=0;s<a;s++)u[s]="String"==n?this.charAt(o+s):this[o+s];return u}});var _a={exports:{}};!function(t){var e=function(t){var e,n=Object.prototype,o=n.hasOwnProperty,i="function"==typeof Symbol?Symbol:{},a=i.iterator||"@@iterator",u=i.asyncIterator||"@@asyncIterator",s=i.toStringTag||"@@toStringTag";function c(t,e,r){return Object.defineProperty(t,e,{value:r,enumerable:!0,configurable:!0,writable:!0}),t[e]}try{c({},"")}catch(t){c=function(t,e,r){return t[e]=r}}function f(t,e,r,n){var o=e&&e.prototype instanceof g?e:g,i=Object.create(o.prototype),a=new P(n||[]);return i._invoke=function(t,e,r){var n=h;return function(o,i){if(n===y)throw new Error("Generator is already running");if(n===d){if("throw"===o)throw i;return L()}for(r.method=o,r.arg=i;;){var a=r.delegate;if(a){var u=A(a,r);if(u){if(u===v)continue;return u}}if("next"===r.method)r.sent=r._sent=r.arg;else if("throw"===r.method){if(n===h)throw n=d,r.arg;r.dispatchException(r.arg)}else"return"===r.method&&r.abrupt("return",r.arg);n=y;var s=l(t,e,r);if("normal"===s.type){if(n=r.done?d:p,s.arg===v)continue;return{value:s.arg,done:r.done}}"throw"===s.type&&(n=d,r.method="throw",r.arg=s.arg)}}}(t,r,a),i}function l(t,e,r){try{return{type:"normal",arg:t.call(e,r)}}catch(t){return{type:"throw",arg:t}}}t.wrap=f;var h="suspendedStart",p="suspendedYield",y="executing",d="completed",v={};function g(){}function b(){}function m(){}var w={};w[a]=function(){return this};var x=Object.getPrototypeOf,_=x&&x(x(I([])));_&&_!==n&&o.call(_,a)&&(w=_);var S=m.prototype=g.prototype=Object.create(w);function E(t){["next","throw","return"].forEach((function(e){c(t,e,(function(t){return this._invoke(e,t)}))}))}function O(t,e){function n(i,a,u,s){var c=l(t[i],t,a);if("throw"!==c.type){var f=c.arg,h=f.value;return h&&"object"===r(h)&&o.call(h,"__await")?e.resolve(h.__await).then((function(t){n("next",t,u,s)}),(function(t){n("throw",t,u,s)})):e.resolve(h).then((function(t){f.value=t,u(f)}),(function(t){return n("throw",t,u,s)}))}s(c.arg)}var i;this._invoke=function(t,r){function o(){return new e((function(e,o){n(t,r,e,o)}))}return i=i?i.then(o,o):o()}}function A(t,r){var n=t.iterator[r.method];if(n===e){if(r.delegate=null,"throw"===r.method){if(t.iterator.return&&(r.method="return",r.arg=e,A(t,r),"throw"===r.method))return v;r.method="throw",r.arg=new TypeError("The iterator does not provide a 'throw' method")}return v}var o=l(n,t.iterator,r.arg);if("throw"===o.type)return r.method="throw",r.arg=o.arg,r.delegate=null,v;var i=o.arg;return i?i.done?(r[t.resultName]=i.value,r.next=t.nextLoc,"return"!==r.method&&(r.method="next",r.arg=e),r.delegate=null,v):i:(r.method="throw",r.arg=new TypeError("iterator result is not an object"),r.delegate=null,v)}function T(t){var e={tryLoc:t[0]};1 in t&&(e.catchLoc=t[1]),2 in t&&(e.finallyLoc=t[2],e.afterLoc=t[3]),this.tryEntries.push(e)}function j(t){var e=t.completion||{};e.type="normal",delete e.arg,t.completion=e}function P(t){this.tryEntries=[{tryLoc:"root"}],t.forEach(T,this),this.reset(!0)}function I(t){if(t){var r=t[a];if(r)return r.call(t);if("function"==typeof t.next)return t;if(!isNaN(t.length)){var n=-1,i=function r(){for(;++n<t.length;)if(o.call(t,n))return r.value=t[n],r.done=!1,r;return r.value=e,r.done=!0,r};return i.next=i}}return{next:L}}function L(){return{value:e,done:!0}}return b.prototype=S.constructor=m,m.constructor=b,b.displayName=c(m,s,"GeneratorFunction"),t.isGeneratorFunction=function(t){var e="function"==typeof t&&t.constructor;return!!e&&(e===b||"GeneratorFunction"===(e.displayName||e.name))},t.mark=function(t){return Object.setPrototypeOf?Object.setPrototypeOf(t,m):(t.__proto__=m,c(t,s,"GeneratorFunction")),t.prototype=Object.create(S),t},t.awrap=function(t){return{__await:t}},E(O.prototype),O.prototype[u]=function(){return this},t.AsyncIterator=O,t.async=function(e,r,n,o,i){void 0===i&&(i=Promise);var a=new O(f(e,r,n,o),i);return t.isGeneratorFunction(r)?a:a.next().then((function(t){return t.done?t.value:a.next()}))},E(S),c(S,s,"Generator"),S[a]=function(){return this},S.toString=function(){return"[object Generator]"},t.keys=function(t){var e=[];for(var r in t)e.push(r);return e.reverse(),function r(){for(;e.length;){var n=e.pop();if(n in t)return r.value=n,r.done=!1,r}return r.done=!0,r}},t.values=I,P.prototype={constructor:P,reset:function(t){if(this.prev=0,this.next=0,this.sent=this._sent=e,this.done=!1,this.delegate=null,this.method="next",this.arg=e,this.tryEntries.forEach(j),!t)for(var r in this)"t"===r.charAt(0)&&o.call(this,r)&&!isNaN(+r.slice(1))&&(this[r]=e)},stop:function(){this.done=!0;var t=this.tryEntries[0].completion;if("throw"===t.type)throw t.arg;return this.rval},dispatchException:function(t){if(this.done)throw t;var r=this;function n(n,o){return u.type="throw",u.arg=t,r.next=n,o&&(r.method="next",r.arg=e),!!o}for(var i=this.tryEntries.length-1;i>=0;--i){var a=this.tryEntries[i],u=a.completion;if("root"===a.tryLoc)return n("end");if(a.tryLoc<=this.prev){var s=o.call(a,"catchLoc"),c=o.call(a,"finallyLoc");if(s&&c){if(this.prev<a.catchLoc)return n(a.catchLoc,!0);if(this.prev<a.finallyLoc)return n(a.finallyLoc)}else if(s){if(this.prev<a.catchLoc)return n(a.catchLoc,!0)}else{if(!c)throw new Error("try statement without catch or finally");if(this.prev<a.finallyLoc)return n(a.finallyLoc)}}}},abrupt:function(t,e){for(var r=this.tryEntries.length-1;r>=0;--r){var n=this.tryEntries[r];if(n.tryLoc<=this.prev&&o.call(n,"finallyLoc")&&this.prev<n.finallyLoc){var i=n;break}}i&&("break"===t||"continue"===t)&&i.tryLoc<=e&&e<=i.finallyLoc&&(i=null);var a=i?i.completion:{};return a.type=t,a.arg=e,i?(this.method="next",this.next=i.finallyLoc,v):this.complete(a)},complete:function(t,e){if("throw"===t.type)throw t.arg;return"break"===t.type||"continue"===t.type?this.next=t.arg:"return"===t.type?(this.rval=this.arg=t.arg,this.method="return",this.next="end"):"normal"===t.type&&e&&(this.next=e),v},finish:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.finallyLoc===t)return this.complete(r.completion,r.afterLoc),j(r),v}},catch:function(t){for(var e=this.tryEntries.length-1;e>=0;--e){var r=this.tryEntries[e];if(r.tryLoc===t){var n=r.completion;if("throw"===n.type){var o=n.arg;j(r)}return o}}throw new Error("illegal catch attempt")},delegateYield:function(t,r,n){return this.delegate={iterator:I(t),resultName:r,nextLoc:n},"next"===this.method&&(this.arg=e),v}},t}(t.exports);try{regeneratorRuntime=e}catch(t){Function("r","regeneratorRuntime = r")(e)}}(_a);var Sa=_a.exports;rt(rt.S,"Number",{MAX_SAFE_INTEGER:9007199254740991});var Ea=h,Oa=y,Aa=function(t,e){if(Oa(t),!Ea(e)&&null!==e)throw TypeError(e+": can't set as prototype!")},Ta={set:Object.setPrototypeOf||("__proto__"in{}?function(t,e,r){try{(r=z(Function.call,Ue.f(Object.prototype,"__proto__").set,2))(t,[]),e=!(t instanceof Array)}catch(t){e=!0}return function(t,n){return Aa(t,n),e?t.__proto__=n:r(t,n),t}}({},!1):void 0),check:Aa},ja=h,Pa=Ta.set,Ia=rt,La=Ft,Fa=u,Ra="\t\n\v\f\r   ᠎             　\u2028\u2029\ufeff",Na="["+Ra+"]",ka=RegExp("^"+Na+Na+"*"),Ba=RegExp(Na+Na+"*$"),Ma=function(t,e,r){var n={},o=Fa((function(){return!!Ra[t]()||"​"!="​"[t]()})),i=n[t]=o?e(Ua):Ra[t];r&&(n[r]=i),Ia(Ia.P+Ia.F*o,"String",n)},Ua=Ma.trim=function(t,e){return t=String(La(t)),1&e&&(t=t.replace(ka,"")),2&e&&(t=t.replace(Ba,"")),t},Ca=Ma,Da=n.exports,Ga=a,qa=Pt,Va=function(t,e,r){var n,o=e.constructor;return o!==r&&"function"==typeof o&&(n=o.prototype)!==r.prototype&&ja(n)&&Pa&&Pa(t,n),t},Wa=x,Ha=u,$a=Le.f,Ya=Ue.f,Ja=l.f,za=Ca.trim,Xa="Number",Ka=Da.Number,Qa=Ka,Za=Ka.prototype,tu=qa(Pe(Za))==Xa,eu="trim"in String.prototype,ru=function(t){var e=Wa(t,!1);if("string"==typeof e&&e.length>2){var r,n,o,i=(e=eu?e.trim():za(e,3)).charCodeAt(0);if(43===i||45===i){if(88===(r=e.charCodeAt(2))||120===r)return NaN}else if(48===i){switch(e.charCodeAt(1)){case 66:case 98:n=2,o=49;break;case 79:case 111:n=8,o=55;break;default:return+e}for(var a,u=e.slice(2),s=0,c=u.length;s<c;s++)if((a=u.charCodeAt(s))<48||a>o)return NaN;return parseInt(u,n)}}return+e};if(!Ka(" 0o1")||!Ka("0b1")||Ka("+0x1")){Ka=function(t){var e=arguments.length<1?0:t,r=this;return r instanceof Ka&&(tu?Ha((function(){Za.valueOf.call(r)})):qa(r)!=Xa)?Va(new Qa(ru(e)),r,Ka):ru(e)};for(var nu,ou=s?$a(Qa):"MAX_VALUE,MIN_VALUE,NaN,NEGATIVE_INFINITY,POSITIVE_INFINITY,EPSILON,isFinite,isInteger,isNaN,isSafeInteger,MAX_SAFE_INTEGER,MIN_SAFE_INTEGER,parseFloat,parseInt,isInteger".split(","),iu=0;ou.length>iu;iu++)Ga(Qa,nu=ou[iu])&&!Ga(Ka,nu)&&Ja(Ka,nu,Ya(Qa,nu));Ka.prototype=Za,Za.constructor=Ka,I.exports(Da,Xa,Ka)}var au,uu,su=h,cu=Pt,fu=pt.exports("match"),lu=Kr(!0),hu=function(t,e,r){return e+(r?lu(t,e).length:1)},pu=An,yu=RegExp.prototype.exec,du=function(t,e){var r=t.exec;if("function"==typeof r){var n=r.call(t,e);if("object"!=typeof n)throw new TypeError("RegExp exec method returned something other than an Object or null");return n}if("RegExp"!==pu(t))throw new TypeError("RegExp#exec called on incompatible receiver");return yu.call(t,e)},vu=ca,gu=RegExp.prototype.exec,bu=String.prototype.replace,mu=gu,wu=(au=/a/,uu=/b*/g,gu.call(au,"a"),gu.call(uu,"a"),0!==au.lastIndex||0!==uu.lastIndex),xu=void 0!==/()??/.exec("")[1];(wu||xu)&&(mu=function(t){var e,r,n,o,i=this;return xu&&(r=new RegExp("^"+i.source+"$(?!\\s)",vu.call(i))),wu&&(e=i.lastIndex),n=gu.call(i,t),wu&&n&&(i.lastIndex=i.global?n.index+n[0].length:e),xu&&n&&n.length>1&&bu.call(n[0],r,(function(){for(o=1;o<arguments.length-2;o++)void 0===arguments[o]&&(n[o]=void 0)})),n});var _u=mu;rt({target:"RegExp",proto:!0,forced:_u!==/./.exec},{exec:_u});var Su=I.exports,Eu=P,Ou=u,Au=Ft,Tu=pt.exports,ju=_u,Pu=Tu("species"),Iu=!Ou((function(){var t=/./;return t.exec=function(){var t=[];return t.groups={a:"7"},t},"7"!=="".replace(t,"$<a>")})),Lu=function(){var t=/(?:)/,e=t.exec;t.exec=function(){return e.apply(this,arguments)};var r="ab".split(t);return 2===r.length&&"a"===r[0]&&"b"===r[1]}(),Fu=function(t,e,r){var n=Tu(t),o=!Ou((function(){var e={};return e[n]=function(){return 7},7!=""[t](e)})),i=o?!Ou((function(){var e=!1,r=/a/;return r.exec=function(){return e=!0,null},"split"===t&&(r.constructor={},r.constructor[Pu]=function(){return r}),r[n](""),!e})):void 0;if(!o||!i||"replace"===t&&!Iu||"split"===t&&!Lu){var a=/./[n],u=r(Au,n,""[t],(function(t,e,r,n,i){return e.exec===ju?o&&!i?{done:!0,value:a.call(e,r,n)}:{done:!0,value:t.call(r,e,n)}:{done:!1}})),s=u[0],c=u[1];Su(String.prototype,t,s),Eu(RegExp.prototype,n,2==e?function(t,e){return c.call(t,this,e)}:function(t){return c.call(t,this)})}},Ru=function(t){var e;return su(t)&&(void 0!==(e=t[fu])?!!e:"RegExp"==cu(t))},Nu=y,ku=Uo,Bu=hu,Mu=Gt,Uu=du,Cu=_u,Du=u,Gu=Math.min,qu=[].push,Vu=4294967295,Wu=!Du((function(){RegExp(Vu,"y")}));Fu("split",2,(function(t,e,r,n){var o;return o="c"=="abbc".split(/(b)*/)[1]||4!="test".split(/(?:)/,-1).length||2!="ab".split(/(?:ab)*/).length||4!=".".split(/(.?)(.?)/).length||".".split(/()()/).length>1||"".split(/.?/).length?function(t,e){var n=String(this);if(void 0===t&&0===e)return[];if(!Ru(t))return r.call(n,t,e);for(var o,i,a,u=[],s=(t.ignoreCase?"i":"")+(t.multiline?"m":"")+(t.unicode?"u":"")+(t.sticky?"y":""),c=0,f=void 0===e?Vu:e>>>0,l=new RegExp(t.source,s+"g");(o=Cu.call(l,n))&&!((i=l.lastIndex)>c&&(u.push(n.slice(c,o.index)),o.length>1&&o.index<n.length&&qu.apply(u,o.slice(1)),a=o[0].length,c=i,u.length>=f));)l.lastIndex===o.index&&l.lastIndex++;return c===n.length?!a&&l.test("")||u.push(""):u.push(n.slice(c)),u.length>f?u.slice(0,f):u}:"0".split(void 0,0).length?function(t,e){return void 0===t&&0===e?[]:r.call(this,t,e)}:r,[function(r,n){var i=t(this),a=null==r?void 0:r[e];return void 0!==a?a.call(r,i,n):o.call(String(i),r,n)},function(t,e){var i=n(o,t,this,e,o!==r);if(i.done)return i.value;var a=Nu(t),u=String(this),s=ku(a,RegExp),c=a.unicode,f=(a.ignoreCase?"i":"")+(a.multiline?"m":"")+(a.unicode?"u":"")+(Wu?"y":"g"),l=new s(Wu?a:"^(?:"+a.source+")",f),h=void 0===e?Vu:e>>>0;if(0===h)return[];if(0===u.length)return null===Uu(l,u)?[u]:[];for(var p=0,y=0,d=[];y<u.length;){l.lastIndex=Wu?y:0;var v,g=Uu(l,Wu?u:u.slice(y));if(null===g||(v=Gu(Mu(l.lastIndex+(Wu?0:y)),u.length))===p)y=Bu(u,y,c);else{if(d.push(u.slice(p,y)),d.length===h)return d;for(var b=1;b<=g.length-1;b++)if(d.push(g[b]),d.length===h)return d;y=p=v}}return d.push(u.slice(p)),d}]}));var Hu=l,$u=A,Yu=z,Ju=rt,zu=ve,Xu=po,Ku=bo,Qu=Gt,Zu=function(t,e,r){e in t?Hu.f(t,e,$u(0,r)):t[e]=r},ts=_o;Ju(Ju.S+Ju.F*!Pi((function(t){Array.from(t)})),"Array",{from:function(t){var e,r,n,o,i=zu(t),a="function"==typeof this?this:Array,u=arguments.length,s=u>1?arguments[1]:void 0,c=void 0!==s,f=0,l=ts(i);if(c&&(s=Yu(s,u>2?arguments[2]:void 0,2)),null==l||a==Array&&Ku(l))for(r=new a(e=Qu(i.length));e>f;f++)Zu(r,f,c?s(i[f],f):i[f]);else for(o=l.call(i),r=new a;!(n=o.next()).done;f++)Zu(r,f,c?Xu(o,s,[n.value,f],!0):n.value);return r.length=f,r}});var es,rs,ns=h,os=ye,is=pt.exports("species"),as=function(t){var e;return os(t)&&("function"!=typeof(e=t.constructor)||e!==Array&&!os(e.prototype)||(e=void 0),ns(e)&&null===(e=e[is])&&(e=void 0)),void 0===e?Array:e},us=z,ss=Lt,cs=ve,fs=Gt,ls=function(t,e){return new(as(t))(e)},hs=function(t,e){var r=1==t,n=2==t,o=3==t,i=4==t,a=6==t,u=5==t||a,s=e||ls;return function(e,c,f){for(var l,h,p=cs(e),y=ss(p),d=us(c,f,3),v=fs(y.length),g=0,b=r?s(e,v):n?s(e,0):void 0;v>g;g++)if((u||g in y)&&(h=d(l=y[g],g,p),t))if(r)b[g]=h;else if(h)switch(t){case 3:return!0;case 5:return l;case 6:return g;case 2:b.push(l)}else if(i)return!1;return a?-1:o||i?i:b}},ps=u,ys=rt,ds=hs(1);ys(ys.P+ys.F*(rs=!0,!((es=[].map)&&ps((function(){rs?es.call(null,(function(){}),1):es.call(null)})))),"Array",{map:function(t){return ds(this,t,arguments[1])}}),no("getOwnPropertyNames",(function(){return Ie.f}));for(var vs,gs={exports:{}},bs=n.exports,ms=P,ws=R,xs=ws("typed_array"),_s=ws("view"),Ss=!(!bs.ArrayBuffer||!bs.DataView),Es=Ss,Os=0,As="Int8Array,Uint8Array,Uint8ClampedArray,Int16Array,Uint16Array,Int32Array,Uint32Array,Float32Array,Float64Array".split(",");Os<9;)(vs=bs[As[Os++]])?(ms(vs.prototype,xs,!0),ms(vs.prototype,_s,!0)):Es=!1;var Ts={ABV:Ss,CONSTR:Es,TYPED:xs,VIEW:_s},js={},Ps=Ut,Is=Gt,Ls=function(t){if(void 0===t)return 0;var e=Ps(t),r=Is(e);if(e!==r)throw RangeError("Wrong length!");return r},Fs=ve,Rs=Ht,Ns=Gt,ks=function(t){for(var e=Fs(this),r=Ns(e.length),n=arguments.length,o=Rs(n>1?arguments[1]:void 0,r),i=n>2?arguments[2]:void 0,a=void 0===i?r:Rs(i,r);a>o;)e[o++]=t;return e};!function(t){var e=n.exports,r=s,o=Ts,i=P,a=vi,c=u,f=fo,h=Ut,p=Gt,y=Ls,d=Le.f,v=l.f,g=ks,b=xt,m="ArrayBuffer",w="DataView",x="Wrong index!",_=e.ArrayBuffer,S=e.DataView,E=e.Math,O=e.RangeError,A=e.Infinity,T=_,j=E.abs,I=E.pow,L=E.floor,F=E.log,R=E.LN2,N="buffer",k="byteLength",B="byteOffset",M=r?"_b":N,U=r?"_l":k,C=r?"_o":B;function D(t,e,r){var n,o,i,a=new Array(r),u=8*r-e-1,s=(1<<u)-1,c=s>>1,f=23===e?I(2,-24)-I(2,-77):0,l=0,h=t<0||0===t&&1/t<0?1:0;for((t=j(t))!=t||t===A?(o=t!=t?1:0,n=s):(n=L(F(t)/R),t*(i=I(2,-n))<1&&(n--,i*=2),(t+=n+c>=1?f/i:f*I(2,1-c))*i>=2&&(n++,i/=2),n+c>=s?(o=0,n=s):n+c>=1?(o=(t*i-1)*I(2,e),n+=c):(o=t*I(2,c-1)*I(2,e),n=0));e>=8;a[l++]=255&o,o/=256,e-=8);for(n=n<<e|o,u+=e;u>0;a[l++]=255&n,n/=256,u-=8);return a[--l]|=128*h,a}function G(t,e,r){var n,o=8*r-e-1,i=(1<<o)-1,a=i>>1,u=o-7,s=r-1,c=t[s--],f=127&c;for(c>>=7;u>0;f=256*f+t[s],s--,u-=8);for(n=f&(1<<-u)-1,f>>=-u,u+=e;u>0;n=256*n+t[s],s--,u-=8);if(0===f)f=1-a;else{if(f===i)return n?NaN:c?-A:A;n+=I(2,e),f-=a}return(c?-1:1)*n*I(2,f-e)}function q(t){return t[3]<<24|t[2]<<16|t[1]<<8|t[0]}function V(t){return[255&t]}function W(t){return[255&t,t>>8&255]}function H(t){return[255&t,t>>8&255,t>>16&255,t>>24&255]}function $(t){return D(t,52,8)}function Y(t){return D(t,23,4)}function J(t,e,r){v(t.prototype,e,{get:function(){return this[r]}})}function z(t,e,r,n){var o=y(+r);if(o+e>t[U])throw O(x);var i=t[M]._b,a=o+t[C],u=i.slice(a,a+e);return n?u:u.reverse()}function X(t,e,r,n,o,i){var a=y(+r);if(a+e>t[U])throw O(x);for(var u=t[M]._b,s=a+t[C],c=n(+o),f=0;f<e;f++)u[s+f]=c[i?f:e-f-1]}if(o.ABV){if(!c((function(){_(1)}))||!c((function(){new _(-1)}))||c((function(){return new _,new _(1.5),new _(NaN),_.name!=m}))){for(var K,Q=(_=function(t){return f(this,_),new T(y(t))}).prototype=T.prototype,Z=d(T),tt=0;Z.length>tt;)(K=Z[tt++])in _||i(_,K,T[K]);Q.constructor=_}var et=new S(new _(2)),rt=S.prototype.setInt8;et.setInt8(0,2147483648),et.setInt8(1,2147483649),!et.getInt8(0)&&et.getInt8(1)||a(S.prototype,{setInt8:function(t,e){rt.call(this,t,e<<24>>24)},setUint8:function(t,e){rt.call(this,t,e<<24>>24)}},!0)}else _=function(t){f(this,_,m);var e=y(t);this._b=g.call(new Array(e),0),this[U]=e},S=function(t,e,r){f(this,S,w),f(t,_,w);var n=t[U],o=h(e);if(o<0||o>n)throw O("Wrong offset!");if(o+(r=void 0===r?n-o:p(r))>n)throw O("Wrong length!");this[M]=t,this[C]=o,this[U]=r},r&&(J(_,k,"_l"),J(S,N,"_b"),J(S,k,"_l"),J(S,B,"_o")),a(S.prototype,{getInt8:function(t){return z(this,1,t)[0]<<24>>24},getUint8:function(t){return z(this,1,t)[0]},getInt16:function(t){var e=z(this,2,t,arguments[1]);return(e[1]<<8|e[0])<<16>>16},getUint16:function(t){var e=z(this,2,t,arguments[1]);return e[1]<<8|e[0]},getInt32:function(t){return q(z(this,4,t,arguments[1]))},getUint32:function(t){return q(z(this,4,t,arguments[1]))>>>0},getFloat32:function(t){return G(z(this,4,t,arguments[1]),23,4)},getFloat64:function(t){return G(z(this,8,t,arguments[1]),52,8)},setInt8:function(t,e){X(this,1,t,V,e)},setUint8:function(t,e){X(this,1,t,V,e)},setInt16:function(t,e){X(this,2,t,W,e,arguments[2])},setUint16:function(t,e){X(this,2,t,W,e,arguments[2])},setInt32:function(t,e){X(this,4,t,H,e,arguments[2])},setUint32:function(t,e){X(this,4,t,H,e,arguments[2])},setFloat32:function(t,e){X(this,4,t,Y,e,arguments[2])},setFloat64:function(t,e){X(this,8,t,$,e,arguments[2])}});b(_,m),b(S,w),i(S.prototype,o.VIEW,!0),t.ArrayBuffer=_,t.DataView=S}(js);var Bs=ve,Ms=Ht,Us=Gt,Cs=[].copyWithin||function(t,e){var r=Bs(this),n=Us(r.length),o=Ms(t,n),i=Ms(e,n),a=arguments.length>2?arguments[2]:void 0,u=Math.min((void 0===a?n:Ms(a,n))-i,n-o),s=1;for(i<o&&o<i+u&&(s=-1,i+=u-1,o+=u-1);u-- >0;)i in r?r[o]=r[i]:delete r[o],o+=s,i+=s;return r};if(s){var Ds=n.exports,Gs=u,qs=rt,Vs=Ts,Ws=js,Hs=z,$s=fo,Ys=A,Js=P,zs=vi,Xs=Ut,Ks=Gt,Qs=Ls,Zs=Ht,tc=x,ec=a,rc=An,nc=h,oc=ve,ic=bo,ac=Pe,uc=sn,sc=Le.f,cc=_o,fc=R,lc=pt.exports,hc=hs,pc=zt,yc=Uo,dc=kn,vc=Qr,gc=Pi,bc=xi,mc=ks,wc=Cs,xc=l,_c=Ue,Sc=xc.f,Ec=_c.f,Oc=Ds.RangeError,Ac=Ds.TypeError,Tc=Ds.Uint8Array,jc="ArrayBuffer",Pc="SharedArrayBuffer",Ic="BYTES_PER_ELEMENT",Lc=Array.prototype,Fc=Ws.ArrayBuffer,Rc=Ws.DataView,Nc=hc(0),kc=hc(2),Bc=hc(3),Mc=hc(4),Uc=hc(5),Cc=hc(6),Dc=pc(!0),Gc=pc(!1),qc=dc.values,Vc=dc.keys,Wc=dc.entries,Hc=Lc.lastIndexOf,$c=Lc.reduce,Yc=Lc.reduceRight,Jc=Lc.join,zc=Lc.sort,Xc=Lc.slice,Kc=Lc.toString,Qc=Lc.toLocaleString,Zc=lc("iterator"),tf=lc("toStringTag"),ef=fc("typed_constructor"),rf=fc("def_constructor"),nf=Vs.CONSTR,of=Vs.TYPED,af=Vs.VIEW,uf="Wrong length!",sf=hc(1,(function(t,e){return pf(yc(t,t[rf]),e)})),cf=Gs((function(){return 1===new Tc(new Uint16Array([1]).buffer)[0]})),ff=!!Tc&&!!Tc.prototype.set&&Gs((function(){new Tc(1).set({})})),lf=function(t,e){var r=Xs(t);if(r<0||r%e)throw Oc("Wrong offset!");return r},hf=function(t){if(nc(t)&&of in t)return t;throw Ac(t+" is not a typed array!")},pf=function(t,e){if(!nc(t)||!(ef in t))throw Ac("It is not a typed array constructor!");return new t(e)},yf=function(t,e){return df(yc(t,t[rf]),e)},df=function(t,e){for(var r=0,n=e.length,o=pf(t,n);n>r;)o[r]=e[r++];return o},vf=function(t,e,r){Sc(t,e,{get:function(){return this._d[r]}})},gf=function(t){var e,r,n,o,i,a,u=oc(t),s=arguments.length,c=s>1?arguments[1]:void 0,f=void 0!==c,l=cc(u);if(null!=l&&!ic(l)){for(a=l.call(u),n=[],e=0;!(i=a.next()).done;e++)n.push(i.value);u=n}for(f&&s>2&&(c=Hs(c,arguments[2],2)),e=0,r=Ks(u.length),o=pf(this,r);r>e;e++)o[e]=f?c(u[e],e):u[e];return o},bf=function(){for(var t=0,e=arguments.length,r=pf(this,e);e>t;)r[t]=arguments[t++];return r},mf=!!Tc&&Gs((function(){Qc.call(new Tc(1))})),wf=function(){return Qc.apply(mf?Xc.call(hf(this)):hf(this),arguments)},xf={copyWithin:function(t,e){return wc.call(hf(this),t,e,arguments.length>2?arguments[2]:void 0)},every:function(t){return Mc(hf(this),t,arguments.length>1?arguments[1]:void 0)},fill:function(t){return mc.apply(hf(this),arguments)},filter:function(t){return yf(this,kc(hf(this),t,arguments.length>1?arguments[1]:void 0))},find:function(t){return Uc(hf(this),t,arguments.length>1?arguments[1]:void 0)},findIndex:function(t){return Cc(hf(this),t,arguments.length>1?arguments[1]:void 0)},forEach:function(t){Nc(hf(this),t,arguments.length>1?arguments[1]:void 0)},indexOf:function(t){return Gc(hf(this),t,arguments.length>1?arguments[1]:void 0)},includes:function(t){return Dc(hf(this),t,arguments.length>1?arguments[1]:void 0)},join:function(t){return Jc.apply(hf(this),arguments)},lastIndexOf:function(t){return Hc.apply(hf(this),arguments)},map:function(t){return sf(hf(this),t,arguments.length>1?arguments[1]:void 0)},reduce:function(t){return $c.apply(hf(this),arguments)},reduceRight:function(t){return Yc.apply(hf(this),arguments)},reverse:function(){for(var t,e=this,r=hf(e).length,n=Math.floor(r/2),o=0;o<n;)t=e[o],e[o++]=e[--r],e[r]=t;return e},some:function(t){return Bc(hf(this),t,arguments.length>1?arguments[1]:void 0)},sort:function(t){return zc.call(hf(this),t)},subarray:function(t,e){var r=hf(this),n=r.length,o=Zs(t,n);return new(yc(r,r[rf]))(r.buffer,r.byteOffset+o*r.BYTES_PER_ELEMENT,Ks((void 0===e?n:Zs(e,n))-o))}},_f=function(t,e){return yf(this,Xc.call(hf(this),t,e))},Sf=function(t){hf(this);var e=lf(arguments[1],1),r=this.length,n=oc(t),o=Ks(n.length),i=0;if(o+e>r)throw Oc(uf);for(;i<o;)this[e+i]=n[i++]},Ef={entries:function(){return Wc.call(hf(this))},keys:function(){return Vc.call(hf(this))},values:function(){return qc.call(hf(this))}},Of=function(t,e){return nc(t)&&t[of]&&"symbol"!=typeof e&&e in t&&String(+e)==String(e)},Af=function(t,e){return Of(t,e=tc(e,!0))?Ys(2,t[e]):Ec(t,e)},Tf=function(t,e,r){return!(Of(t,e=tc(e,!0))&&nc(r)&&ec(r,"value"))||ec(r,"get")||ec(r,"set")||r.configurable||ec(r,"writable")&&!r.writable||ec(r,"enumerable")&&!r.enumerable?Sc(t,e,r):(t[e]=r.value,t)};nf||(_c.f=Af,xc.f=Tf),qs(qs.S+qs.F*!nf,"Object",{getOwnPropertyDescriptor:Af,defineProperty:Tf}),Gs((function(){Kc.call({})}))&&(Kc=Qc=function(){return Jc.call(this)});var jf=zs({},xf);zs(jf,Ef),Js(jf,Zc,Ef.values),zs(jf,{slice:_f,set:Sf,constructor:function(){},toString:Kc,toLocaleString:wf}),vf(jf,"buffer","b"),vf(jf,"byteOffset","o"),vf(jf,"byteLength","l"),vf(jf,"length","e"),Sc(jf,tf,{get:function(){return this[of]}}),gs.exports=function(t,e,r,n){var o=t+((n=!!n)?"Clamped":"")+"Array",i="get"+t,a="set"+t,u=Ds[o],s=u||{},c=u&&uc(u),f=!u||!Vs.ABV,l={},h=u&&u.prototype,p=function(t,r){Sc(t,r,{get:function(){return function(t,r){var n=t._d;return n.v[i](r*e+n.o,cf)}(this,r)},set:function(t){return function(t,r,o){var i=t._d;n&&(o=(o=Math.round(o))<0?0:o>255?255:255&o),i.v[a](r*e+i.o,o,cf)}(this,r,t)},enumerable:!0})};f?(u=r((function(t,r,n,i){$s(t,u,o,"_d");var a,s,c,f,l=0,h=0;if(nc(r)){if(!(r instanceof Fc||(f=rc(r))==jc||f==Pc))return of in r?df(u,r):gf.call(u,r);a=r,h=lf(n,e);var y=r.byteLength;if(void 0===i){if(y%e)throw Oc(uf);if((s=y-h)<0)throw Oc(uf)}else if((s=Ks(i)*e)+h>y)throw Oc(uf);c=s/e}else c=Qs(r),a=new Fc(s=c*e);for(Js(t,"_d",{b:a,o:h,l:s,e:c,v:new Rc(a)});l<c;)p(t,l++)})),h=u.prototype=ac(jf),Js(h,"constructor",u)):Gs((function(){u(1)}))&&Gs((function(){new u(-1)}))&&gc((function(t){new u,new u(null),new u(1.5),new u(t)}),!0)||(u=r((function(t,r,n,i){var a;return $s(t,u,o),nc(r)?r instanceof Fc||(a=rc(r))==jc||a==Pc?void 0!==i?new s(r,lf(n,e),i):void 0!==n?new s(r,lf(n,e)):new s(r):of in r?df(u,r):gf.call(u,r):new s(Qs(r))})),Nc(c!==Function.prototype?sc(s).concat(sc(c)):sc(s),(function(t){t in u||Js(u,t,s[t])})),u.prototype=h,h.constructor=u);var y=h[Zc],d=!!y&&("values"==y.name||null==y.name),v=Ef.values;Js(u,ef,!0),Js(h,of,o),Js(h,af,!0),Js(h,rf,u),(n?new u(1)[tf]==o:tf in h)||Sc(h,tf,{get:function(){return o}}),l[o]=u,qs(qs.G+qs.W+qs.F*(u!=s),l),qs(qs.S,o,{BYTES_PER_ELEMENT:e}),qs(qs.S+qs.F*Gs((function(){s.of.call(u,1)})),o,{from:gf,of:bf}),Ic in h||Js(h,Ic,e),qs(qs.P,o,xf),bc(o),qs(qs.P+qs.F*ff,o,{set:Sf}),qs(qs.P+qs.F*!d,o,Ef),h.toString!=Kc&&(h.toString=Kc),qs(qs.P+qs.F*Gs((function(){new u(1).slice()})),o,{slice:_f}),qs(qs.P+qs.F*(Gs((function(){return[1,2].toLocaleString()!=new u([1,2]).toLocaleString()}))||!Gs((function(){h.toLocaleString.call([1,2])}))),o,{toLocaleString:wf}),vc[o]=d?y:v,d||Js(h,Zc,v)}}else gs.exports=function(){};gs.exports("Uint8",1,(function(t){return function(e,r,n){return t(this,e,r,n)}}));var Pf=y,If=ve,Lf=Gt,Ff=Ut,Rf=hu,Nf=du,kf=Math.max,Bf=Math.min,Mf=Math.floor,Uf=/\$([$&`']|\d\d?|<[^>]*>)/g,Cf=/\$([$&`']|\d\d?)/g;Fu("replace",2,(function(t,e,r,n){return[function(n,o){var i=t(this),a=null==n?void 0:n[e];return void 0!==a?a.call(n,i,o):r.call(String(i),n,o)},function(t,e){var i=n(r,t,this,e);if(i.done)return i.value;var a=Pf(t),u=String(this),s="function"==typeof e;s||(e=String(e));var c=a.global;if(c){var f=a.unicode;a.lastIndex=0}for(var l=[];;){var h=Nf(a,u);if(null===h)break;if(l.push(h),!c)break;""===String(h[0])&&(a.lastIndex=Rf(u,Lf(a.lastIndex),f))}for(var p,y="",d=0,v=0;v<l.length;v++){h=l[v];for(var g=String(h[0]),b=kf(Bf(Ff(h.index),u.length),0),m=[],w=1;w<h.length;w++)m.push(void 0===(p=h[w])?p:String(p));var x=h.groups;if(s){var _=[g].concat(m,b,u);void 0!==x&&_.push(x);var S=String(e.apply(void 0,_))}else S=o(g,u,b,m,x,e);b>=d&&(y+=u.slice(d,b)+S,d=b+g.length)}return y+u.slice(d)}];function o(t,e,n,o,i,a){var u=n+t.length,s=o.length,c=Cf;return void 0!==i&&(i=If(i),c=Uf),r.call(a,c,(function(r,a){var c;switch(a.charAt(0)){case"$":return"$";case"&":return t;case"`":return e.slice(0,n);case"'":return e.slice(u);case"<":c=i[a.slice(1,-1)];break;default:var f=+a;if(0===f)return r;if(f>s){var l=Mf(f/10);return 0===l?r:l<=s?void 0===o[l-1]?a.charAt(1):o[l-1]+a.charAt(1):r}c=o[f-1]}return void 0===c?"":c}))}}));var Df="undefined"!=typeof globalThis&&globalThis||"undefined"!=typeof self&&self||void 0!==Df&&Df,Gf="URLSearchParams"in Df,qf="Symbol"in Df&&"iterator"in Symbol,Vf="FileReader"in Df&&"Blob"in Df&&function(){try{return new Blob,!0}catch(t){return!1}}(),Wf="FormData"in Df,Hf="ArrayBuffer"in Df;if(Hf)var $f=["[object Int8Array]","[object Uint8Array]","[object Uint8ClampedArray]","[object Int16Array]","[object Uint16Array]","[object Int32Array]","[object Uint32Array]","[object Float32Array]","[object Float64Array]"],Yf=ArrayBuffer.isView||function(t){return t&&$f.indexOf(Object.prototype.toString.call(t))>-1};function Jf(t){if("string"!=typeof t&&(t=String(t)),/[^a-z0-9\-#$%&'*+.^_`|~!]/i.test(t)||""===t)throw new TypeError('Invalid character in header field name: "'+t+'"');return t.toLowerCase()}function zf(t){return"string"!=typeof t&&(t=String(t)),t}function Xf(t){var e={next:function(){var e=t.shift();return{done:void 0===e,value:e}}};return qf&&(e[Symbol.iterator]=function(){return e}),e}function Kf(t){this.map={},t instanceof Kf?t.forEach((function(t,e){this.append(e,t)}),this):Array.isArray(t)?t.forEach((function(t){this.append(t[0],t[1])}),this):t&&Object.getOwnPropertyNames(t).forEach((function(e){this.append(e,t[e])}),this)}function Qf(t){if(t.bodyUsed)return Promise.reject(new TypeError("Already read"));t.bodyUsed=!0}function Zf(t){return new Promise((function(e,r){t.onload=function(){e(t.result)},t.onerror=function(){r(t.error)}}))}function tl(t){var e=new FileReader,r=Zf(e);return e.readAsArrayBuffer(t),r}function el(t){if(t.slice)return t.slice(0);var e=new Uint8Array(t.byteLength);return e.set(new Uint8Array(t)),e.buffer}function rl(){return this.bodyUsed=!1,this._initBody=function(t){var e;this.bodyUsed=this.bodyUsed,this._bodyInit=t,t?"string"==typeof t?this._bodyText=t:Vf&&Blob.prototype.isPrototypeOf(t)?this._bodyBlob=t:Wf&&FormData.prototype.isPrototypeOf(t)?this._bodyFormData=t:Gf&&URLSearchParams.prototype.isPrototypeOf(t)?this._bodyText=t.toString():Hf&&Vf&&((e=t)&&DataView.prototype.isPrototypeOf(e))?(this._bodyArrayBuffer=el(t.buffer),this._bodyInit=new Blob([this._bodyArrayBuffer])):Hf&&(ArrayBuffer.prototype.isPrototypeOf(t)||Yf(t))?this._bodyArrayBuffer=el(t):this._bodyText=t=Object.prototype.toString.call(t):this._bodyText="",this.headers.get("content-type")||("string"==typeof t?this.headers.set("content-type","text/plain;charset=UTF-8"):this._bodyBlob&&this._bodyBlob.type?this.headers.set("content-type",this._bodyBlob.type):Gf&&URLSearchParams.prototype.isPrototypeOf(t)&&this.headers.set("content-type","application/x-www-form-urlencoded;charset=UTF-8"))},Vf&&(this.blob=function(){var t=Qf(this);if(t)return t;if(this._bodyBlob)return Promise.resolve(this._bodyBlob);if(this._bodyArrayBuffer)return Promise.resolve(new Blob([this._bodyArrayBuffer]));if(this._bodyFormData)throw new Error("could not read FormData body as blob");return Promise.resolve(new Blob([this._bodyText]))},this.arrayBuffer=function(){if(this._bodyArrayBuffer){var t=Qf(this);return t||(ArrayBuffer.isView(this._bodyArrayBuffer)?Promise.resolve(this._bodyArrayBuffer.buffer.slice(this._bodyArrayBuffer.byteOffset,this._bodyArrayBuffer.byteOffset+this._bodyArrayBuffer.byteLength)):Promise.resolve(this._bodyArrayBuffer))}return this.blob().then(tl)}),this.text=function(){var t,e,r,n=Qf(this);if(n)return n;if(this._bodyBlob)return t=this._bodyBlob,e=new FileReader,r=Zf(e),e.readAsText(t),r;if(this._bodyArrayBuffer)return Promise.resolve(function(t){for(var e=new Uint8Array(t),r=new Array(e.length),n=0;n<e.length;n++)r[n]=String.fromCharCode(e[n]);return r.join("")}(this._bodyArrayBuffer));if(this._bodyFormData)throw new Error("could not read FormData body as text");return Promise.resolve(this._bodyText)},Wf&&(this.formData=function(){return this.text().then(il)}),this.json=function(){return this.text().then(JSON.parse)},this}Kf.prototype.append=function(t,e){t=Jf(t),e=zf(e);var r=this.map[t];this.map[t]=r?r+", "+e:e},Kf.prototype.delete=function(t){delete this.map[Jf(t)]},Kf.prototype.get=function(t){return t=Jf(t),this.has(t)?this.map[t]:null},Kf.prototype.has=function(t){return this.map.hasOwnProperty(Jf(t))},Kf.prototype.set=function(t,e){this.map[Jf(t)]=zf(e)},Kf.prototype.forEach=function(t,e){for(var r in this.map)this.map.hasOwnProperty(r)&&t.call(e,this.map[r],r,this)},Kf.prototype.keys=function(){var t=[];return this.forEach((function(e,r){t.push(r)})),Xf(t)},Kf.prototype.values=function(){var t=[];return this.forEach((function(e){t.push(e)})),Xf(t)},Kf.prototype.entries=function(){var t=[];return this.forEach((function(e,r){t.push([r,e])})),Xf(t)},qf&&(Kf.prototype[Symbol.iterator]=Kf.prototype.entries);var nl=["DELETE","GET","HEAD","OPTIONS","POST","PUT"];function ol(t,e){if(!(this instanceof ol))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');var r=(e=e||{}).body;if(t instanceof ol){if(t.bodyUsed)throw new TypeError("Already read");this.url=t.url,this.credentials=t.credentials,e.headers||(this.headers=new Kf(t.headers)),this.method=t.method,this.mode=t.mode,this.signal=t.signal,r||null==t._bodyInit||(r=t._bodyInit,t.bodyUsed=!0)}else this.url=String(t);if(this.credentials=e.credentials||this.credentials||"same-origin",!e.headers&&this.headers||(this.headers=new Kf(e.headers)),this.method=function(t){var e=t.toUpperCase();return nl.indexOf(e)>-1?e:t}(e.method||this.method||"GET"),this.mode=e.mode||this.mode||null,this.signal=e.signal||this.signal,this.referrer=null,("GET"===this.method||"HEAD"===this.method)&&r)throw new TypeError("Body not allowed for GET or HEAD requests");if(this._initBody(r),!("GET"!==this.method&&"HEAD"!==this.method||"no-store"!==e.cache&&"no-cache"!==e.cache)){var n=/([?&])_=[^&]*/;if(n.test(this.url))this.url=this.url.replace(n,"$1_="+(new Date).getTime());else{this.url+=(/\?/.test(this.url)?"&":"?")+"_="+(new Date).getTime()}}}function il(t){var e=new FormData;return t.trim().split("&").forEach((function(t){if(t){var r=t.split("="),n=r.shift().replace(/\+/g," "),o=r.join("=").replace(/\+/g," ");e.append(decodeURIComponent(n),decodeURIComponent(o))}})),e}function al(t,e){if(!(this instanceof al))throw new TypeError('Please use the "new" operator, this DOM object constructor cannot be called as a function.');e||(e={}),this.type="default",this.status=void 0===e.status?200:e.status,this.ok=this.status>=200&&this.status<300,this.statusText=void 0===e.statusText?"":""+e.statusText,this.headers=new Kf(e.headers),this.url=e.url||"",this._initBody(t)}ol.prototype.clone=function(){return new ol(this,{body:this._bodyInit})},rl.call(ol.prototype),rl.call(al.prototype),al.prototype.clone=function(){return new al(this._bodyInit,{status:this.status,statusText:this.statusText,headers:new Kf(this.headers),url:this.url})},al.error=function(){var t=new al(null,{status:0,statusText:""});return t.type="error",t};var ul=[301,302,303,307,308];al.redirect=function(t,e){if(-1===ul.indexOf(e))throw new RangeError("Invalid status code");return new al(null,{status:e,headers:{location:t}})};var sl,cl,fl=Df.DOMException;try{new fl}catch(t){(fl=function(t,e){this.message=t,this.name=e;var r=Error(t);this.stack=r.stack}).prototype=Object.create(Error.prototype),fl.prototype.constructor=fl}function ll(t,e){return new Promise((function(n,o){var i=new ol(t,e);if(i.signal&&i.signal.aborted)return o(new fl("Aborted","AbortError"));var a=new XMLHttpRequest;function u(){a.abort()}a.onload=function(){var t,e,r={status:a.status,statusText:a.statusText,headers:(t=a.getAllResponseHeaders()||"",e=new Kf,t.replace(/\r?\n[\t ]+/g," ").split("\r").map((function(t){return 0===t.indexOf("\n")?t.substr(1,t.length):t})).forEach((function(t){var r=t.split(":"),n=r.shift().trim();if(n){var o=r.join(":").trim();e.append(n,o)}})),e)};r.url="responseURL"in a?a.responseURL:r.headers.get("X-Request-URL");var o="response"in a?a.response:a.responseText;setTimeout((function(){n(new al(o,r))}),0)},a.onerror=function(){setTimeout((function(){o(new TypeError("Network request failed"))}),0)},a.ontimeout=function(){setTimeout((function(){o(new TypeError("Network request failed"))}),0)},a.onabort=function(){setTimeout((function(){o(new fl("Aborted","AbortError"))}),0)},a.open(i.method,function(t){try{return""===t&&Df.location.href?Df.location.href:t}catch(e){return t}}(i.url),!0),"include"===i.credentials?a.withCredentials=!0:"omit"===i.credentials&&(a.withCredentials=!1),"responseType"in a&&(Vf?a.responseType="blob":Hf&&i.headers.get("Content-Type")&&-1!==i.headers.get("Content-Type").indexOf("application/octet-stream")&&(a.responseType="arraybuffer")),!e||"object"!==r(e.headers)||e.headers instanceof Kf?i.headers.forEach((function(t,e){a.setRequestHeader(e,t)})):Object.getOwnPropertyNames(e.headers).forEach((function(t){a.setRequestHeader(t,zf(e.headers[t]))})),i.signal&&(i.signal.addEventListener("abort",u),a.onreadystatechange=function(){4===a.readyState&&i.signal.removeEventListener("abort",u)}),a.send(void 0===i._bodyInit?null:i._bodyInit)}))}function hl(t,e){var r="undefined"!=typeof Symbol&&t[Symbol.iterator]||t["@@iterator"];if(!r){if(Array.isArray(t)||(r=function(t,e){if(!t)return;if("string"==typeof t)return pl(t,e);var r=Object.prototype.toString.call(t).slice(8,-1);"Object"===r&&t.constructor&&(r=t.constructor.name);if("Map"===r||"Set"===r)return Array.from(t);if("Arguments"===r||/^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(r))return pl(t,e)}(t))||e&&t&&"number"==typeof t.length){r&&(t=r);var n=0,o=function(){};return{s:o,n:function(){return n>=t.length?{done:!0}:{done:!1,value:t[n++]}},e:function(t){throw t},f:o}}throw new TypeError("Invalid attempt to iterate non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.")}var i,a=!0,u=!1;return{s:function(){r=r.call(t)},n:function(){var t=r.next();return a=t.done,t},e:function(t){u=!0,i=t},f:function(){try{a||null==r.return||r.return()}finally{if(u)throw i}}}}function pl(t,e){(null==e||e>t.length)&&(e=t.length);for(var r=0,n=new Array(e);r<e;r++)n[r]=t[r];return n}ll.polyfill=!0,Df.fetch||(Df.fetch=ll,Df.Headers=Kf,Df.Request=ol,Df.Response=al),"object"!==r(window.ui)?((sl={}).resolveStorage={},sl.requestCounter=0,sl.initStarted=!1,sl.initFinished=!1,sl.initFailed=!1,window.ui=sl,window.ui.copyright="© Copyright 2021, by Marco Mengelkoch"):sl=window.ui,"object"!==r(window.backend)?(cl={},window.backend=cl):cl=window.backend;return sl.createRequestId=function(){return sl.requestCounter>=Number.MAX_SAFE_INTEGER-1&&(sl.requestCounter=0),sl.requestCounter++},sl.alert=function(t){void 0===window.nimview?alert(t):"function"==typeof window.nimview.alert&&window.nimview.alert(t)},sl.callRequest=function(){var t=e(Sa.mark((function t(e,r,n){var o,i,a,u,s,c;return Sa.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(void 0===r&&(r=""),void 0===n&&(n=[]),void 0!==window.nimview){t.next=12;break}return o=JSON.stringify({request:e,data:n}),i={method:"POST",mode:"cors",cache:"no-cache",headers:{"Content-Type":"application/json"},body:o},a=e,sl.globalToken&&sl.globalToken.length>0&&(i.headers["global-token"]=sl.globalToken),n.length!=Math.min(r.length,r.split(",").length)&&console&&console.log&&console.log(0&e&r&"' data: '"&JSON.stringify(n)&"'"),t.abrupt("return",fetch("/"+a,i).then((function(t){if(t){var e=t.headers.get("global-token");e&&e.length>0&&(sl.globalToken=e,sl.lastToken=Date.now())}return t.text()})));case 12:if("function"!=typeof window.nimview.call){t.next=17;break}return u=sl.createRequestId(),s=JSON.stringify({request:e,data:n,requestId:u}),c=new Promise((function(t,e){sl.resolveStorage[u]=[t,e];var r=window.nimview.call(s);void 0!==r&&(t(r),delete sl.resolveStorage[u])})),t.abrupt("return",c);case 17:case"end":return t.stop()}}),t)})));return function(e,r,n){return t.apply(this,arguments)}}(),sl.addRequest=function(t){"string"==typeof t&&(t=[t,""]);var r=t[0]+"",n=t[1]+"";cl[r]=e(Sa.mark((function t(){var e,o,i,a=arguments;return Sa.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:for(e=a.length,o=new Array(e),i=0;i<e;i++)o[i]=a[i];return t.abrupt("return",sl.callRequest(r,n,o));case 2:case"end":return t.stop()}}),t)})))},sl.applyResponse=function(t,e){void 0!==sl.resolveStorage[t]?(sl.resolveStorage[t][0](e),delete sl.resolveStorage[t]):sl.alert("request id '"+t+"' not found")},sl.rejectResponse=function(t){void 0!==sl.resolveStorage[t]&&(sl.resolveStorage[t][1](),delete sl.resolveStorage[t])},sl.initRequests=e(Sa.mark((function t(){return Sa.wrap((function(t){for(;;)switch(t.prev=t.next){case 0:if(0!=sl.initStarted){t.next=6;break}return sl.initStarted=!0,t.next=4,sl.callRequest("getGlobalToken","",[]).then((function(t){JSON.parse(t).useGlobalToken&&window.setInterval(sl.getGlobalToken,6e4)})).catch((function(){sl.alert("getGlobalToken failed"),sl.initFailed=!0}));case 4:return t.next=6,sl.callRequest("getRequests","",[]).then((function(t){var e,r=hl(JSON.parse(t));try{for(r.s();!(e=r.n()).done;){var n=e.value;sl.addRequest(n)}}catch(t){r.e(t)}finally{r.f()}sl.initFinished=!0})).catch((function(){sl.alert("getRequests failed"),sl.initFailed=!0}));case 6:case"end":return t.stop()}}),t)}))),cl.waitInit=function(){return new Promise((function(t,e){setTimeout((function r(){return sl.initFinished?t():sl.initFailed?e():void setTimeout(r,5)}),1)}))},window.setTimeout(sl.initRequests,3),cl}();
+//# sourceMappingURL=nimview.iife.js.map
